session: split the god module along its seams
session.rs becomes session/ with config (env vocabulary + BrainConfig), recovery (crash
projections + redelivery), storage_state (upload/publish/delete machine), view (protocol
mapping) and engine/{subagents,storage,sandbox} (the three closed capabilities) as
grep-addressable files; mod.rs drops from 10,892 to 6,883 lines. Code moves verbatim;
visibility widened to pub(super) only where the parent calls in.
