+++ crates/brain/src/session/mod.rs
@@ -44,702 +44,25 @@ use std::time::Duration;
 use tokio::sync::{Notify, OnceCell, Semaphore, mpsc, oneshot};
 use tokio_util::sync::CancellationToken;
 
-pub const PROVIDER_HEADER_TIMEOUT_ENV: &str = "BRAIN_PROVIDER_HEADER_TIMEOUT_MS";
-pub const PROVIDER_IDLE_TIMEOUT_ENV: &str = "BRAIN_PROVIDER_IDLE_TIMEOUT_MS";
-pub const PROVIDER_TOTAL_TIMEOUT_ENV: &str = "BRAIN_PROVIDER_TOTAL_TIMEOUT_MS";
-pub const EXTERNAL_TOOL_TIMEOUT_ENV: &str = "BRAIN_EXTERNAL_TOOL_TIMEOUT_MS";
-pub const MAX_MODEL_ROUNDS_ENV: &str = "BRAIN_MAX_MODEL_ROUNDS";
-pub const DEFAULT_MAX_ROUNDS_PER_TURN_ENV: &str = "BRAIN_DEFAULT_MAX_ROUNDS_PER_TURN";
-pub const DRAIN_TIMEOUT_ENV: &str = "BRAIN_DRAIN_TIMEOUT_MS";
-pub const DEFAULT_DRAIN_TIMEOUT_MS: u64 = 110_000;
-pub const MAX_TURNS_ENV: &str = "BRAIN_MAX_TURNS";
-pub const MAX_CONCURRENT_CREATES_ENV: &str = "BRAIN_MAX_CONCURRENT_CREATES";
-pub const MAX_EVENT_FOLLOWERS_ENV: &str = "BRAIN_MAX_EVENT_FOLLOWERS";
-pub const MAX_RESIDENT_SESSIONS_ENV: &str = "BRAIN_MAX_RESIDENT_SESSIONS";
-pub const IDLE_DISCARD_SECONDS_ENV: &str = "BRAIN_IDLE_DISCARD_SECONDS";
-pub const OUTBOUND_ALLOW_PRIVATE_ENV: &str = "BRAIN_OUTBOUND_ALLOW_PRIVATE";
-pub const STORAGE_MAX_OBJECT_BYTES_ENV: &str = "BRAIN_STORAGE_MAX_OBJECT_BYTES";
-pub const STORAGE_MAX_SESSION_BYTES_ENV: &str = "BRAIN_STORAGE_MAX_SESSION_BYTES";
-pub const STORAGE_MAX_TENANT_BYTES_ENV: &str = "BRAIN_STORAGE_MAX_TENANT_BYTES";
-pub const STORAGE_TRANSFER_TTL_ENV: &str = "BRAIN_STORAGE_TRANSFER_TTL_MS";
-pub const MAX_ADDITIONAL_SANDBOXES_ENV: &str = "BRAIN_MAX_ADDITIONAL_SANDBOXES_PER_ROOT";
-pub const EXTERNAL_EXECUTOR_URL_ENV: &str = "BRAIN_EXTERNAL_TOOL_EXECUTOR_URL";
-pub const EXTERNAL_EXECUTOR_TOKEN_ENV: &str = "BRAIN_EXTERNAL_TOOL_EXECUTOR_TOKEN";
-pub const EXTERNAL_EXECUTOR_CAPABILITIES_ENV: &str = "BRAIN_EXTERNAL_TOOL_CAPABILITIES";
-pub const RECOVERY_POLL_ENV: &str = "BRAIN_RECOVERY_POLL_MS";
-pub const RECOVERY_SHARDS_PER_POLL_ENV: &str = "BRAIN_RECOVERY_SHARDS_PER_POLL";
-pub const RECOVERY_PAGE_SIZE_ENV: &str = "BRAIN_RECOVERY_PAGE_SIZE";
-pub const MAX_CONCURRENT_RECOVERIES_ENV: &str = "BRAIN_MAX_CONCURRENT_RECOVERIES";
+mod view;
+pub use view::*;
 
-pub const DEFAULT_PROVIDER_HEADER_TIMEOUT_MS: u64 = 30_000;
-pub const DEFAULT_PROVIDER_IDLE_TIMEOUT_MS: u64 = 60_000;
-pub const DEFAULT_PROVIDER_TOTAL_TIMEOUT_MS: u64 = 15 * 60_000;
-pub const DEFAULT_EXTERNAL_TOOL_TIMEOUT_MS: u64 = 30_000;
-pub const DEFAULT_MAX_CONCURRENT_MODEL_ROUNDS: usize = 64;
-pub const DEFAULT_MAX_CONCURRENT_TURNS: usize = 64;
-pub const DEFAULT_MAX_CONCURRENT_CREATES: usize = 4;
-pub const DEFAULT_MAX_EVENT_FOLLOWERS: usize = 64;
-pub const DEFAULT_MAX_RESIDENT_SESSIONS: usize = 128;
-pub const DEFAULT_IDLE_DISCARD_SECONDS: u64 = 900;
-pub const DEFAULT_STORAGE_MAX_TENANT_BYTES: u64 = 10 * 1024 * 1024 * 1024;
-pub const DEFAULT_MAX_ADDITIONAL_SANDBOXES: usize = 2;
-pub const DEFAULT_RECOVERY_POLL_MS: u64 = 1_000;
-pub const DEFAULT_RECOVERY_SHARDS_PER_POLL: usize = 4;
-pub const DEFAULT_RECOVERY_PAGE_SIZE: usize = 32;
-pub const DEFAULT_MAX_CONCURRENT_RECOVERIES: usize = 16;
+mod storage_state;
+use storage_state::*;
 
-pub const MIN_PROVIDER_HEADER_TIMEOUT_MS: u64 = 1_000;
-pub const MAX_PROVIDER_HEADER_TIMEOUT_MS: u64 = 5 * 60_000;
-pub const MIN_PROVIDER_IDLE_TIMEOUT_MS: u64 = 1_000;
-pub const MAX_PROVIDER_IDLE_TIMEOUT_MS: u64 = 5 * 60_000;
-pub const MIN_PROVIDER_TOTAL_TIMEOUT_MS: u64 = 1_000;
-pub const MAX_PROVIDER_TOTAL_TIMEOUT_MS: u64 = 60 * 60_000;
-pub const MIN_EXTERNAL_TOOL_TIMEOUT_MS: u64 = 1_000;
-pub const MAX_EXTERNAL_TOOL_TIMEOUT_MS: u64 = 5 * 60_000;
-pub const MAX_CONCURRENT_MODEL_ROUNDS: usize = 1_024;
-pub const MAX_CONCURRENT_TURNS: usize = 1_024;
-pub const MAX_CONCURRENT_CREATES: usize = 64;
-pub const MAX_EVENT_FOLLOWERS: usize = 1_024;
-pub const MAX_RESIDENT_SESSIONS: usize = 4_096;
-pub const MAX_IDLE_DISCARD_SECONDS: u64 = 24 * 60 * 60;
-pub const MAX_STORAGE_POLICY_BYTES: u64 = i64::MAX as u64;
-pub const MIN_STORAGE_TRANSFER_TTL_MS: u64 = 60_000;
-pub const MAX_STORAGE_TRANSFER_TTL_MS: u64 = 24 * 60 * 60 * 1_000;
-pub const MAX_ADDITIONAL_SANDBOXES: usize = 64;
-pub const MIN_RECOVERY_POLL_MS: u64 = 1;
-pub const MAX_RECOVERY_POLL_MS: u64 = 60_000;
-pub const MAX_RECOVERY_PAGE_SIZE: usize = 100;
-pub const MAX_CONCURRENT_RECOVERIES: usize = 1_024;
-pub const MAX_EXTERNAL_EXECUTOR_URL_BYTES: usize = 2_048;
-pub const MAX_EXTERNAL_EXECUTOR_TOKEN_BYTES: usize = 8 * 1_024;
-pub const MAX_EXTERNAL_EXECUTOR_CAPABILITIES: usize = 128;
-pub const MAX_EXTERNAL_EXECUTOR_CAPABILITY_BYTES: usize = 128;
+#[path = "engine/sandbox.rs"]
+mod engine_sandbox;
 
-/// Process configuration: the knobs that are NOT adapters.
-#[derive(Debug, Clone)]
-pub struct BrainConfig {
-    /// Admission: concurrent model rounds across the process.
-    pub max_concurrent_model_rounds: usize,
-    /// Admission: concurrent active turns across the process.
-    pub max_concurrent_turns: usize,
-    /// Expensive create resolution/bundle validation admission. HTTP compositions should also
-    /// acquire before buffering; this core guard protects direct/local callers.
-    pub max_concurrent_creates: usize,
-    /// Process-wide SSE follower admission. Each admitted distinct session may retain one small
-    /// fixed live-event ring; slow consumers disconnect and replay durable records.
-    pub max_event_followers: usize,
-    /// Hard process-wide bound on resident actors/folds. Admission is non-blocking; a saturated
-    /// process returns backpressure and durable sessions transparently hydrate after an idle
-    /// resident releases its slot.
-    pub max_resident_sessions: usize,
-    /// Idle residency before the actor discards its fold and exits.
-    pub idle_discard: Duration,
-    pub context_soft_tokens: usize,
-    pub context_hard_tokens: usize,
-    pub context_tail_tokens: usize,
-    /// Provider response header/first-byte, inter-event idle, and total-call budgets.
-    pub provider_header_timeout: Duration,
-    pub provider_idle_timeout: Duration,
-    pub provider_total_timeout: Duration,
-    /// Sealed per-turn model-round ceiling for new sessions. Kernel runaway authorization, not
-    /// working policy: the graceful closing round wraps the turn up in text at the cap.
-    pub default_max_rounds: u32,
-    /// How long a draining process waits for admitted turns before exiting anyway. Size it
-    /// under the orchestrator's stop timeout (Fargate caps stopTimeout at 120 s).
-    pub drain_timeout: Duration,
-    /// Whether user-controlled provider base URLs may reach loopback/private/link-local addresses.
-    /// Production keeps this false; explicit local development may opt into private endpoints.
-    pub outbound_allow_private: bool,
-    /// Maximum bytes buffered by one public file upload or download.
-    /// Sealed durable-storage limits. These are host policy, copied into each session's immutable
-    /// configuration so a later deployment cannot silently widen an existing capability.
-    pub storage_max_object_bytes: u64,
-    pub storage_max_session_bytes: u64,
-    /// Authoritative aggregate ceiling across published and reserved bytes for every session
-    /// owned by one tenant. Journal adapters enforce this in the same atomic decision as HEAD.
-    pub storage_max_tenant_bytes: u64,
-    /// Append-only journal retention is independently bounded from model-context compaction.
-    /// These limits include roots, ordinary descendants, ended sessions, and charged recovery
-    /// headroom until explicit physical deletion releases the identity and bytes.
-    pub journal_max_session_bytes: u64,
-    pub journal_max_tenant_bytes: u64,
-    pub journal_max_tenant_sessions: u64,
-    pub storage_transfer_ttl: Duration,
-    /// Host-sealed root-tree quota for simultaneously live additional sandboxes. The shared
-    /// default target is excluded; terminated IDs remain tombstoned but release a slot.
-    pub max_additional_sandboxes_per_root: u32,
-    /// Optional host executor shared by every external tool declaration. The URL and service
-    /// credential are process configuration and never enter the sealed model prefix.
-    pub external_executor_url: Option<String>,
-    pub external_executor_token: Option<ProviderKey>,
-    /// Stable capabilities registered by the HTTP executor. An empty set advertises none.
-    pub external_executor_capabilities: HashSet<String>,
-    /// Trusted host registry for fixed official capability IDs. Public Tool JSON may reference a
-    /// capability but cannot choose scope, completion, replay policy, or input ceilings.
-    pub official_capabilities: HashMap<String, crate::config::ServerToolPolicy>,
-    pub external_call_timeout: Duration,
-    /// Background recovery discovery. The index is eventually consistent; correctness comes
-    /// from the base-journal claim/fence before an actor is started.
-    pub recovery_poll_interval: Duration,
-    pub recovery_shards_per_poll: usize,
-    pub recovery_page_size: usize,
-    /// Independent cap for background effect recovery so a due-time wave cannot create an
-    /// unbounded resident/adapter fan-out.
-    pub max_concurrent_recoveries: usize,
-}
+#[path = "engine/storage.rs"]
+mod engine_storage;
 
-impl Default for BrainConfig {
-    fn default() -> Self {
-        Self {
-            max_concurrent_model_rounds: DEFAULT_MAX_CONCURRENT_MODEL_ROUNDS,
-            max_concurrent_turns: DEFAULT_MAX_CONCURRENT_TURNS,
-            max_concurrent_creates: DEFAULT_MAX_CONCURRENT_CREATES,
-            max_event_followers: DEFAULT_MAX_EVENT_FOLLOWERS,
-            max_resident_sessions: DEFAULT_MAX_RESIDENT_SESSIONS,
-            idle_discard: Duration::from_secs(DEFAULT_IDLE_DISCARD_SECONDS),
-            context_soft_tokens: DEFAULT_CONTEXT_SOFT_TOKENS,
-            context_hard_tokens: DEFAULT_CONTEXT_HARD_TOKENS,
-            context_tail_tokens: DEFAULT_CONTEXT_TAIL_TOKENS,
-            provider_header_timeout: Duration::from_millis(DEFAULT_PROVIDER_HEADER_TIMEOUT_MS),
-            provider_idle_timeout: Duration::from_millis(DEFAULT_PROVIDER_IDLE_TIMEOUT_MS),
-            provider_total_timeout: Duration::from_millis(DEFAULT_PROVIDER_TOTAL_TIMEOUT_MS),
-            default_max_rounds: crate::config::Limits::default().max_rounds,
-            drain_timeout: Duration::from_millis(DEFAULT_DRAIN_TIMEOUT_MS),
-            outbound_allow_private: false,
-            storage_max_object_bytes: crate::storage::DEFAULT_MAX_STORAGE_OBJECT_BYTES,
-            storage_max_session_bytes: crate::storage::DEFAULT_MAX_SESSION_STORAGE_BYTES,
-            storage_max_tenant_bytes: DEFAULT_STORAGE_MAX_TENANT_BYTES,
-            journal_max_session_bytes: crate::journal::DEFAULT_MAX_SESSION_JOURNAL_BYTES,
-            journal_max_tenant_bytes: crate::journal::DEFAULT_MAX_TENANT_JOURNAL_BYTES,
-            journal_max_tenant_sessions: crate::journal::DEFAULT_MAX_TENANT_RETAINED_SESSIONS,
-            storage_transfer_ttl: Duration::from_millis(
-                crate::storage::DEFAULT_STORAGE_TRANSFER_TTL_MS,
-            ),
-            max_additional_sandboxes_per_root: DEFAULT_MAX_ADDITIONAL_SANDBOXES as u32,
-            external_executor_url: None,
-            external_executor_token: None,
-            external_executor_capabilities: HashSet::new(),
-            official_capabilities: HashMap::new(),
-            external_call_timeout: Duration::from_millis(DEFAULT_EXTERNAL_TOOL_TIMEOUT_MS),
-            recovery_poll_interval: Duration::from_millis(DEFAULT_RECOVERY_POLL_MS),
-            recovery_shards_per_poll: DEFAULT_RECOVERY_SHARDS_PER_POLL,
-            recovery_page_size: DEFAULT_RECOVERY_PAGE_SIZE,
-            max_concurrent_recoveries: DEFAULT_MAX_CONCURRENT_RECOVERIES,
-        }
-    }
-}
+#[path = "engine/subagents.rs"]
+mod engine_subagents;
 
-impl BrainConfig {
-    /// Load every Brain-owned process policy without silently replacing or clamping malformed
-    /// operator input. `Default` is deliberately environment-independent for tests and explicit
-    /// local composition; hosted processes must call this constructor.
-    pub fn from_env() -> Result<Self> {
-        Self::from_env_values(|name| {
-            let Some(raw) = std::env::var_os(name) else {
-                return Ok(None);
-            };
-            raw.into_string()
-                .map(Some)
-                .map_err(|_| BrainError::Invalid(format!("{name} must contain valid UTF-8 text")))
-        })
-    }
-
-    #[allow(clippy::field_reassign_with_default)]
-    fn from_env_values(mut read: impl FnMut(&str) -> Result<Option<String>>) -> Result<Self> {
-        let mut cfg = Self::default();
-        cfg.max_concurrent_model_rounds = parse_env_usize(
-            MAX_MODEL_ROUNDS_ENV,
-            read(MAX_MODEL_ROUNDS_ENV)?.as_deref(),
-            DEFAULT_MAX_CONCURRENT_MODEL_ROUNDS,
-            1,
-            MAX_CONCURRENT_MODEL_ROUNDS,
-        )?;
-        cfg.max_concurrent_turns = parse_env_usize(
-            MAX_TURNS_ENV,
-            read(MAX_TURNS_ENV)?.as_deref(),
-            DEFAULT_MAX_CONCURRENT_TURNS,
-            1,
-            MAX_CONCURRENT_TURNS,
-        )?;
-        cfg.max_concurrent_creates = parse_env_usize(
-            MAX_CONCURRENT_CREATES_ENV,
-            read(MAX_CONCURRENT_CREATES_ENV)?.as_deref(),
-            DEFAULT_MAX_CONCURRENT_CREATES,
-            1,
-            MAX_CONCURRENT_CREATES,
-        )?;
-        cfg.max_event_followers = parse_env_usize(
-            MAX_EVENT_FOLLOWERS_ENV,
-            read(MAX_EVENT_FOLLOWERS_ENV)?.as_deref(),
-            DEFAULT_MAX_EVENT_FOLLOWERS,
-            1,
-            MAX_EVENT_FOLLOWERS,
-        )?;
-        cfg.max_resident_sessions = parse_env_usize(
-            MAX_RESIDENT_SESSIONS_ENV,
-            read(MAX_RESIDENT_SESSIONS_ENV)?.as_deref(),
-            DEFAULT_MAX_RESIDENT_SESSIONS,
-            1,
-            MAX_RESIDENT_SESSIONS,
-        )?;
-        cfg.idle_discard = Duration::from_secs(parse_env_u64(
-            IDLE_DISCARD_SECONDS_ENV,
-            read(IDLE_DISCARD_SECONDS_ENV)?.as_deref(),
-            DEFAULT_IDLE_DISCARD_SECONDS,
-            1,
-            MAX_IDLE_DISCARD_SECONDS,
-        )?);
-        cfg.provider_header_timeout = Duration::from_millis(parse_env_u64(
-            PROVIDER_HEADER_TIMEOUT_ENV,
-            read(PROVIDER_HEADER_TIMEOUT_ENV)?.as_deref(),
-            DEFAULT_PROVIDER_HEADER_TIMEOUT_MS,
-            MIN_PROVIDER_HEADER_TIMEOUT_MS,
-            MAX_PROVIDER_HEADER_TIMEOUT_MS,
-        )?);
-        cfg.provider_idle_timeout = Duration::from_millis(parse_env_u64(
-            PROVIDER_IDLE_TIMEOUT_ENV,
-            read(PROVIDER_IDLE_TIMEOUT_ENV)?.as_deref(),
-            DEFAULT_PROVIDER_IDLE_TIMEOUT_MS,
-            MIN_PROVIDER_IDLE_TIMEOUT_MS,
-            MAX_PROVIDER_IDLE_TIMEOUT_MS,
-        )?);
-        cfg.provider_total_timeout = Duration::from_millis(parse_env_u64(
-            PROVIDER_TOTAL_TIMEOUT_ENV,
-            read(PROVIDER_TOTAL_TIMEOUT_ENV)?.as_deref(),
-            DEFAULT_PROVIDER_TOTAL_TIMEOUT_MS,
-            MIN_PROVIDER_TOTAL_TIMEOUT_MS,
-            MAX_PROVIDER_TOTAL_TIMEOUT_MS,
-        )?);
-        cfg.external_call_timeout = Duration::from_millis(parse_env_u64(
-            EXTERNAL_TOOL_TIMEOUT_ENV,
-            read(EXTERNAL_TOOL_TIMEOUT_ENV)?.as_deref(),
-            DEFAULT_EXTERNAL_TOOL_TIMEOUT_MS,
-            MIN_EXTERNAL_TOOL_TIMEOUT_MS,
-            MAX_EXTERNAL_TOOL_TIMEOUT_MS,
-        )?);
-        cfg.default_max_rounds = u32::try_from(parse_env_u64(
-            DEFAULT_MAX_ROUNDS_PER_TURN_ENV,
-            read(DEFAULT_MAX_ROUNDS_PER_TURN_ENV)?.as_deref(),
-            crate::config::Limits::default().max_rounds as u64,
-            1,
-            100_000,
-        )?)
-        .expect("bounded above by 100000");
-        cfg.drain_timeout = Duration::from_millis(parse_env_u64(
-            DRAIN_TIMEOUT_ENV,
-            read(DRAIN_TIMEOUT_ENV)?.as_deref(),
-            DEFAULT_DRAIN_TIMEOUT_MS,
-            1_000,
-            600_000,
-        )?);
-        cfg.outbound_allow_private = parse_env_bool(
-            OUTBOUND_ALLOW_PRIVATE_ENV,
-            read(OUTBOUND_ALLOW_PRIVATE_ENV)?.as_deref(),
-            false,
-        )?;
-        cfg.storage_max_object_bytes = parse_env_u64(
-            STORAGE_MAX_OBJECT_BYTES_ENV,
-            read(STORAGE_MAX_OBJECT_BYTES_ENV)?.as_deref(),
-            crate::storage::DEFAULT_MAX_STORAGE_OBJECT_BYTES,
-            1,
-            MAX_STORAGE_POLICY_BYTES,
-        )?;
-        cfg.storage_max_session_bytes = parse_env_u64(
-            STORAGE_MAX_SESSION_BYTES_ENV,
-            read(STORAGE_MAX_SESSION_BYTES_ENV)?.as_deref(),
-            crate::storage::DEFAULT_MAX_SESSION_STORAGE_BYTES,
-            1,
-            MAX_STORAGE_POLICY_BYTES,
-        )?;
-        cfg.storage_max_tenant_bytes = parse_env_u64(
-            STORAGE_MAX_TENANT_BYTES_ENV,
-            read(STORAGE_MAX_TENANT_BYTES_ENV)?.as_deref(),
-            DEFAULT_STORAGE_MAX_TENANT_BYTES,
-            1,
-            MAX_STORAGE_POLICY_BYTES,
-        )?;
-        cfg.journal_max_session_bytes = parse_env_u64(
-            crate::journal::JOURNAL_MAX_SESSION_BYTES_ENV,
-            read(crate::journal::JOURNAL_MAX_SESSION_BYTES_ENV)?.as_deref(),
-            crate::journal::DEFAULT_MAX_SESSION_JOURNAL_BYTES,
-            crate::journal::MIN_SESSION_JOURNAL_BYTES,
-            crate::journal::MAX_JOURNAL_BYTES,
-        )?;
-        cfg.journal_max_tenant_bytes = parse_env_u64(
-            crate::journal::JOURNAL_MAX_TENANT_BYTES_ENV,
-            read(crate::journal::JOURNAL_MAX_TENANT_BYTES_ENV)?.as_deref(),
-            crate::journal::DEFAULT_MAX_TENANT_JOURNAL_BYTES,
-            crate::journal::MIN_SESSION_JOURNAL_BYTES,
-            crate::journal::MAX_JOURNAL_BYTES,
-        )?;
-        cfg.journal_max_tenant_sessions = parse_env_u64(
-            crate::journal::JOURNAL_MAX_TENANT_SESSIONS_ENV,
-            read(crate::journal::JOURNAL_MAX_TENANT_SESSIONS_ENV)?.as_deref(),
-            crate::journal::DEFAULT_MAX_TENANT_RETAINED_SESSIONS,
-            crate::journal::MIN_TENANT_RETAINED_SESSIONS,
-            crate::journal::MAX_TENANT_RETAINED_SESSIONS,
-        )?;
-        cfg.storage_transfer_ttl = Duration::from_millis(parse_env_u64(
-            STORAGE_TRANSFER_TTL_ENV,
-            read(STORAGE_TRANSFER_TTL_ENV)?.as_deref(),
-            crate::storage::DEFAULT_STORAGE_TRANSFER_TTL_MS,
-            MIN_STORAGE_TRANSFER_TTL_MS,
-            MAX_STORAGE_TRANSFER_TTL_MS,
-        )?);
-        cfg.max_additional_sandboxes_per_root = u32::try_from(parse_env_usize(
-            MAX_ADDITIONAL_SANDBOXES_ENV,
-            read(MAX_ADDITIONAL_SANDBOXES_ENV)?.as_deref(),
-            DEFAULT_MAX_ADDITIONAL_SANDBOXES,
-            1,
-            MAX_ADDITIONAL_SANDBOXES,
-        )?)
-        .expect("the validated additional-sandbox maximum fits u32");
-        cfg.external_executor_url = parse_optional_env_string(
-            EXTERNAL_EXECUTOR_URL_ENV,
-            read(EXTERNAL_EXECUTOR_URL_ENV)?,
-            MAX_EXTERNAL_EXECUTOR_URL_BYTES,
-        )?;
-        cfg.external_executor_token = parse_optional_env_string(
-            EXTERNAL_EXECUTOR_TOKEN_ENV,
-            read(EXTERNAL_EXECUTOR_TOKEN_ENV)?,
-            MAX_EXTERNAL_EXECUTOR_TOKEN_BYTES,
-        )?
-        .map(ProviderKey::new);
-        cfg.external_executor_capabilities =
-            parse_capabilities(read(EXTERNAL_EXECUTOR_CAPABILITIES_ENV)?)?;
-        cfg.recovery_poll_interval = Duration::from_millis(parse_env_u64(
-            RECOVERY_POLL_ENV,
-            read(RECOVERY_POLL_ENV)?.as_deref(),
-            DEFAULT_RECOVERY_POLL_MS,
-            MIN_RECOVERY_POLL_MS,
-            MAX_RECOVERY_POLL_MS,
-        )?);
-        cfg.recovery_shards_per_poll = parse_env_usize(
-            RECOVERY_SHARDS_PER_POLL_ENV,
-            read(RECOVERY_SHARDS_PER_POLL_ENV)?.as_deref(),
-            DEFAULT_RECOVERY_SHARDS_PER_POLL,
-            1,
-            crate::journal::RECOVERY_SHARDS,
-        )?;
-        cfg.recovery_page_size = parse_env_usize(
-            RECOVERY_PAGE_SIZE_ENV,
-            read(RECOVERY_PAGE_SIZE_ENV)?.as_deref(),
-            DEFAULT_RECOVERY_PAGE_SIZE,
-            1,
-            MAX_RECOVERY_PAGE_SIZE,
-        )?;
-        cfg.max_concurrent_recoveries = parse_env_usize(
-            MAX_CONCURRENT_RECOVERIES_ENV,
-            read(MAX_CONCURRENT_RECOVERIES_ENV)?.as_deref(),
-            DEFAULT_MAX_CONCURRENT_RECOVERIES,
-            1,
-            MAX_CONCURRENT_RECOVERIES,
-        )?;
-        cfg.validate()?;
-        Ok(cfg)
-    }
-
-    pub fn validate(&self) -> Result<()> {
-        validate_usize_range(
-            MAX_MODEL_ROUNDS_ENV,
-            self.max_concurrent_model_rounds,
-            1,
-            MAX_CONCURRENT_MODEL_ROUNDS,
-        )?;
-        validate_usize_range(
-            MAX_TURNS_ENV,
-            self.max_concurrent_turns,
-            1,
-            MAX_CONCURRENT_TURNS,
-        )?;
-        validate_usize_range(
-            MAX_CONCURRENT_CREATES_ENV,
-            self.max_concurrent_creates,
-            1,
-            MAX_CONCURRENT_CREATES,
-        )?;
-        validate_usize_range(
-            MAX_EVENT_FOLLOWERS_ENV,
-            self.max_event_followers,
-            1,
-            MAX_EVENT_FOLLOWERS,
-        )?;
-        validate_usize_range(
-            MAX_RESIDENT_SESSIONS_ENV,
-            self.max_resident_sessions,
-            1,
-            MAX_RESIDENT_SESSIONS,
-        )?;
-        validate_timeout(
-            IDLE_DISCARD_SECONDS_ENV,
-            self.idle_discard,
-            1,
-            MAX_IDLE_DISCARD_SECONDS.saturating_mul(1_000),
-        )?;
-        if self.context_tail_tokens == 0
-            || self.context_tail_tokens >= self.context_soft_tokens
-            || self.context_soft_tokens >= self.context_hard_tokens
-        {
-            return Err(BrainError::Invalid(
-                "context tail must be positive and ordered tail < soft < hard".into(),
-            ));
-        }
-        crate::journal::JournalRetentionLimits {
-            session_bytes: self.journal_max_session_bytes,
-            tenant_bytes: self.journal_max_tenant_bytes,
-            tenant_sessions: self.journal_max_tenant_sessions,
-        }
-        .validate()?;
-        validate_timeout(
-            PROVIDER_HEADER_TIMEOUT_ENV,
-            self.provider_header_timeout,
-            MIN_PROVIDER_HEADER_TIMEOUT_MS,
-            MAX_PROVIDER_HEADER_TIMEOUT_MS,
-        )?;
-        validate_timeout(
-            PROVIDER_IDLE_TIMEOUT_ENV,
-            self.provider_idle_timeout,
-            MIN_PROVIDER_IDLE_TIMEOUT_MS,
-            MAX_PROVIDER_IDLE_TIMEOUT_MS,
-        )?;
-        validate_timeout(
-            PROVIDER_TOTAL_TIMEOUT_ENV,
-            self.provider_total_timeout,
-            MIN_PROVIDER_TOTAL_TIMEOUT_MS,
-            MAX_PROVIDER_TOTAL_TIMEOUT_MS,
-        )?;
-        validate_timeout(
-            EXTERNAL_TOOL_TIMEOUT_ENV,
-            self.external_call_timeout,
-            MIN_EXTERNAL_TOOL_TIMEOUT_MS,
-            MAX_EXTERNAL_TOOL_TIMEOUT_MS,
-        )?;
-        if self.provider_total_timeout < self.provider_header_timeout
-            || self.provider_total_timeout < self.provider_idle_timeout
-        {
-            return Err(BrainError::Invalid(format!(
-                "{PROVIDER_TOTAL_TIMEOUT_ENV} must be greater than or equal to both {PROVIDER_HEADER_TIMEOUT_ENV} and {PROVIDER_IDLE_TIMEOUT_ENV}"
-            )));
-        }
-        if self.storage_max_object_bytes == 0
-            || self.storage_max_object_bytes > MAX_STORAGE_POLICY_BYTES
-            || self.storage_max_session_bytes > MAX_STORAGE_POLICY_BYTES
-            || self.storage_max_tenant_bytes > MAX_STORAGE_POLICY_BYTES
-            || self.storage_max_object_bytes > self.storage_max_session_bytes
-            || self.storage_max_session_bytes > self.storage_max_tenant_bytes
-        {
-            return Err(BrainError::Invalid(format!(
-                "storage byte policies must be positive, representable, and ordered {STORAGE_MAX_OBJECT_BYTES_ENV} <= {STORAGE_MAX_SESSION_BYTES_ENV} <= {STORAGE_MAX_TENANT_BYTES_ENV}"
-            )));
-        }
-        validate_timeout(
-            STORAGE_TRANSFER_TTL_ENV,
-            self.storage_transfer_ttl,
-            MIN_STORAGE_TRANSFER_TTL_MS,
-            MAX_STORAGE_TRANSFER_TTL_MS,
-        )?;
-        validate_usize_range(
-            MAX_ADDITIONAL_SANDBOXES_ENV,
-            self.max_additional_sandboxes_per_root as usize,
-            1,
-            MAX_ADDITIONAL_SANDBOXES,
-        )?;
-        validate_timeout(
-            RECOVERY_POLL_ENV,
-            self.recovery_poll_interval,
-            MIN_RECOVERY_POLL_MS,
-            MAX_RECOVERY_POLL_MS,
-        )?;
-        validate_usize_range(
-            RECOVERY_SHARDS_PER_POLL_ENV,
-            self.recovery_shards_per_poll,
-            1,
-            crate::journal::RECOVERY_SHARDS,
-        )?;
-        validate_usize_range(
-            RECOVERY_PAGE_SIZE_ENV,
-            self.recovery_page_size,
-            1,
-            MAX_RECOVERY_PAGE_SIZE,
-        )?;
-        validate_usize_range(
-            MAX_CONCURRENT_RECOVERIES_ENV,
-            self.max_concurrent_recoveries,
-            1,
-            MAX_CONCURRENT_RECOVERIES,
-        )?;
-        validate_external_executor_config(self)?;
-        Ok(())
-    }
-}
-
-fn validate_external_executor_config(cfg: &BrainConfig) -> Result<()> {
-    if cfg.external_executor_url.is_none()
-        && (cfg.external_executor_token.is_some() || !cfg.external_executor_capabilities.is_empty())
-    {
-        return Err(BrainError::Invalid(format!(
-            "{EXTERNAL_EXECUTOR_TOKEN_ENV} and {EXTERNAL_EXECUTOR_CAPABILITIES_ENV} require {EXTERNAL_EXECUTOR_URL_ENV}"
-        )));
-    }
-    if cfg.external_executor_capabilities.len() > MAX_EXTERNAL_EXECUTOR_CAPABILITIES
-        || cfg.external_executor_capabilities.iter().any(|capability| {
-            capability.is_empty()
-                || capability.len() > MAX_EXTERNAL_EXECUTOR_CAPABILITY_BYTES
-                || !capability
-                    .bytes()
-                    .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'.' | b'_' | b'-'))
-        })
-    {
-        return Err(BrainError::Invalid(format!(
-            "{EXTERNAL_EXECUTOR_CAPABILITIES_ENV} exceeds its count, byte, or identifier bound"
-        )));
-    }
-    if let Some(endpoint) = &cfg.external_executor_url {
-        crate::external::HttpExternalToolExecutor::new(
-            endpoint.clone(),
-            cfg.external_executor_token
-                .as_ref()
-                .map(|token| token.expose().to_owned()),
-            cfg.external_call_timeout,
-            cfg.external_executor_capabilities.iter().cloned(),
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_timeout(name: &str, value: Duration, minimum_ms: u64, maximum_ms: u64) -> Result<()> {
-    let millis = u64::try_from(value.as_millis()).map_err(|_| {
-        BrainError::Invalid(format!(
-            "{name} must be between {minimum_ms} and {maximum_ms} milliseconds"
-        ))
-    })?;
-    if !(minimum_ms..=maximum_ms).contains(&millis) {
-        return Err(BrainError::Invalid(format!(
-            "{name} must be between {minimum_ms} and {maximum_ms} milliseconds"
-        )));
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-fn parse_strict_env_u64(k: &str, raw: Option<&str>, default: u64) -> Result<u64> {
-    parse_env_u64(k, raw, default, 0, u64::MAX)
-}
-
-fn parse_env_u64(
-    name: &str,
-    raw: Option<&str>,
-    default: u64,
-    minimum: u64,
-    maximum: u64,
-) -> Result<u64> {
-    let Some(raw) = raw else {
-        return Ok(default);
-    };
-    let value = raw.parse::<u64>().map_err(|_| {
-        BrainError::Invalid(format!("{name} must contain an unsigned decimal integer"))
-    })?;
-    if !(minimum..=maximum).contains(&value) {
-        return Err(BrainError::Invalid(format!(
-            "{name} must be between {minimum} and {maximum}"
-        )));
-    }
-    Ok(value)
-}
-
-fn parse_env_usize(
-    name: &str,
-    raw: Option<&str>,
-    default: usize,
-    minimum: usize,
-    maximum: usize,
-) -> Result<usize> {
-    let value = parse_env_u64(name, raw, default as u64, minimum as u64, maximum as u64)?;
-    usize::try_from(value)
-        .map_err(|_| BrainError::Invalid(format!("{name} must be between {minimum} and {maximum}")))
-}
-
-fn parse_env_bool(name: &str, raw: Option<&str>, default: bool) -> Result<bool> {
-    match raw {
-        None => Ok(default),
-        Some("true") => Ok(true),
-        Some("false") => Ok(false),
-        Some(_) => Err(BrainError::Invalid(format!(
-            "{name} must be exactly true or false"
-        ))),
-    }
-}
-
-fn parse_optional_env_string(
-    name: &str,
-    raw: Option<String>,
-    maximum_bytes: usize,
-) -> Result<Option<String>> {
-    let Some(raw) = raw else {
-        return Ok(None);
-    };
-    if raw.is_empty() || raw.len() > maximum_bytes {
-        return Err(BrainError::Invalid(format!(
-            "{name} must contain between 1 and {maximum_bytes} UTF-8 bytes"
-        )));
-    }
-    Ok(Some(raw))
-}
-
-fn parse_capabilities(raw: Option<String>) -> Result<HashSet<String>> {
-    let Some(raw) = raw else {
-        return Ok(HashSet::new());
-    };
-    if raw.is_empty() {
-        return Err(BrainError::Invalid(format!(
-            "{EXTERNAL_EXECUTOR_CAPABILITIES_ENV} must not be empty when set"
-        )));
-    }
-    let mut capabilities = HashSet::new();
-    for capability in raw.split(',').map(str::trim) {
-        if capability.is_empty()
-            || capability.len() > MAX_EXTERNAL_EXECUTOR_CAPABILITY_BYTES
-            || !capability
-                .bytes()
-                .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'.' | b'_' | b'-'))
-            || !capabilities.insert(capability.to_owned())
-        {
-            return Err(BrainError::Invalid(format!(
-                "{EXTERNAL_EXECUTOR_CAPABILITIES_ENV} contains an invalid or duplicate capability"
-            )));
-        }
-    }
-    if capabilities.len() > MAX_EXTERNAL_EXECUTOR_CAPABILITIES {
-        return Err(BrainError::Invalid(format!(
-            "{EXTERNAL_EXECUTOR_CAPABILITIES_ENV} contains more than {MAX_EXTERNAL_EXECUTOR_CAPABILITIES} capabilities"
-        )));
-    }
-    Ok(capabilities)
-}
-
-fn validate_usize_range(name: &str, value: usize, minimum: usize, maximum: usize) -> Result<()> {
-    if !(minimum..=maximum).contains(&value) {
-        return Err(BrainError::Invalid(format!(
-            "{name} must be between {minimum} and {maximum}"
-        )));
-    }
-    Ok(())
-}
-
-/// Process-wide reclaim policy. Measurements showed that allocators retain dropped session memory
-/// unless reclamation is requested explicitly.
-fn reclaim_policy() -> &'static crate::reclaim::ReclaimPolicy {
-    static POLICY: std::sync::OnceLock<crate::reclaim::ReclaimPolicy> = std::sync::OnceLock::new();
-    POLICY
-        .get_or_init(|| crate::reclaim::ReclaimPolicy::new(crate::reclaim::DEFAULT_THRESHOLD_BYTES))
-}
-
-/// How turns obtain a provider. Overridable so tests can inject the scripted fake.
-pub type ProviderFactory = Arc<dyn Fn(Dialect) -> Arc<dyn Provider> + Send + Sync>;
-
-fn default_provider_factory() -> ProviderFactory {
-    Arc::new(|d| Arc::from(crate::provider::for_dialect(d)))
-}
+mod config;
+pub use config::*;
+mod recovery;
+use recovery::*;
 
 /// The supervisor: the composed parts and the resident-session map.
 pub struct Brain {
@@ -3207,374 +2530,6 @@ impl Brain {
         }
     }
 
-    /// Engine-owned model facade over the same ordinary child-session resources used by the
-    /// public API. Model-visible names never select an implementation; the sealed capability ID
-    /// reaches this method only after the ToolCall intent is durable.
-    pub(crate) async fn execute_child_capability(
-        self: &Arc<Self>,
-        parent_id: &str,
-        operation_id: &str,
-        input: serde_json::Value,
-        cancel: CancellationToken,
-    ) -> CallOutcome {
-        let started = std::time::Instant::now();
-        let action = input.get("action").and_then(serde_json::Value::as_str);
-        let result: Result<serde_json::Value> = async {
-            if cancel.is_cancelled() {
-                return Err(BrainError::Cancelled);
-            }
-            match action {
-                Some("spawn_agent") => {
-                    let task_name = required_child_string(&input, "task_name")?;
-                    let message = required_child_string(&input, "message")?;
-                    let fork_turns = input
-                        .get("fork_turns")
-                        .and_then(serde_json::Value::as_str)
-                        .map(str::to_owned);
-                    let child = self
-                        .create_child(
-                            parent_id,
-                            message,
-                            Some(task_name),
-                            fork_turns,
-                            Some(operation_id),
-                        )
-                        .await?;
-                    Ok(serde_json::to_value(child)?)
-                }
-                Some("send_message" | "follow_up") => {
-                    let child_id = required_child_string(&input, "child_id")?;
-                    let message = required_child_string(&input, "message")?;
-                    self.get_child(parent_id, &child_id).await?;
-                    let content =
-                        MessageRequestContent::String(message.parse().map_err(|error| {
-                            BrainError::Invalid(format!("child message: {error}"))
-                        })?);
-                    let (turn_id, seq) = self
-                        .message_with_metadata_idempotent(
-                            &child_id,
-                            content,
-                            HashMap::new(),
-                            Some(operation_id),
-                        )
-                        .await?;
-                    Ok(serde_json::json!({
-                        "child_id": child_id,
-                        "turn_id": turn_id,
-                        "seq": seq,
-                    }))
-                }
-                Some("peek") => {
-                    let child_id = required_child_string(&input, "child_id")?;
-                    Ok(serde_json::to_value(
-                        self.get_child(parent_id, &child_id).await?,
-                    )?)
-                }
-                Some("wait") => {
-                    let child_id = required_child_string(&input, "child_id")?;
-                    let timeout = input
-                        .get("timeout_ms")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(30_000)
-                        .min(300_000);
-                    let wait =
-                        self.wait_child(parent_id, &child_id, Duration::from_millis(timeout));
-                    let child = tokio::select! {
-                        result = wait => result?,
-                        _ = cancel.cancelled() => return Err(BrainError::Cancelled),
-                    };
-                    Ok(serde_json::to_value(child)?)
-                }
-                Some("list_children") => {
-                    let cursor = input.get("cursor").and_then(serde_json::Value::as_str);
-                    let limit = input
-                        .get("limit")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(20)
-                        .min(100) as usize;
-                    let (data, next_cursor) = self.list_children(parent_id, cursor, limit).await?;
-                    Ok(serde_json::json!({
-                        "has_more": next_cursor.is_some(),
-                        "next_cursor": next_cursor,
-                        "data": data,
-                    }))
-                }
-                Some("interrupt_agent") => {
-                    let child_id = required_child_string(&input, "child_id")?;
-                    self.get_child(parent_id, &child_id).await?;
-                    Ok(serde_json::to_value(self.cancel(&child_id).await?)?)
-                }
-                Some("end_agent") => {
-                    let child_id = required_child_string(&input, "child_id")?;
-                    self.get_child(parent_id, &child_id).await?;
-                    Ok(serde_json::to_value(self.end(&child_id).await?)?)
-                }
-                Some(other) => Err(BrainError::Invalid(format!(
-                    "unknown subagents action {other:?}"
-                ))),
-                None => Err(BrainError::Invalid("subagents action is required".into())),
-            }
-        }
-        .await;
-        match result {
-            Ok(value) => CallOutcome {
-                outcome: TerminalOutcome::Completed,
-                content: serde_json::to_string(&value)
-                    .unwrap_or_else(|_| "child operation completed".into()),
-                value: Some(value),
-                is_error: false,
-                exit_code: None,
-                duration_ms: started.elapsed().as_millis() as u64,
-                truncated: false,
-                terminal: None,
-            },
-            Err(BrainError::Cancelled) => CallOutcome {
-                outcome: TerminalOutcome::Cancelled,
-                content: "child operation cancelled".into(),
-                value: None,
-                is_error: true,
-                exit_code: None,
-                duration_ms: started.elapsed().as_millis() as u64,
-                truncated: false,
-                terminal: None,
-            },
-            Err(error) => {
-                let mut outcome = CallOutcome::failed(error.to_string());
-                outcome.duration_ms = started.elapsed().as_millis() as u64;
-                outcome
-            }
-        }
-    }
-
-    /// Engine-owned durable-storage facade. The caller passes the turn's already-claimed mutable
-    /// state so storage reservations and their ToolCall live under one fence; this must never
-    /// recurse through the session actor while that same turn is running.
-    pub(crate) async fn execute_storage_capability(
-        self: &Arc<Self>,
-        session_id: &str,
-        operation_id: &str,
-        input: serde_json::Value,
-        cancel: CancellationToken,
-        st: &mut TurnState,
-    ) -> Result<CallOutcome> {
-        let started = std::time::Instant::now();
-        let operation = async {
-            let action = input
-                .get("action")
-                .and_then(serde_json::Value::as_str)
-                .ok_or_else(|| BrainError::Invalid("storage action is required".into()))?;
-            match action {
-                "list" => {
-                    let prefix = optional_bounded_string(&input, "prefix", 1024)?;
-                    if let Some(prefix) = &prefix {
-                        validate_storage_prefix(prefix)?;
-                    }
-                    let cursor = optional_bounded_string(&input, "cursor", 4096)?;
-                    let limit = input
-                        .get("limit")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(20);
-                    if !(1..=100).contains(&limit) {
-                        return Err(BrainError::Invalid(
-                            "storage list limit must be between 1 and 100".into(),
-                        ));
-                    }
-                    ensure_storage_readable(&st.head, session_id)?;
-                    let page = self
-                        .storage_port()?
-                        .list(
-                            session_id,
-                            prefix.as_deref(),
-                            cursor.as_deref(),
-                            limit as u32,
-                        )
-                        .await?;
-                    Ok(serde_json::to_value(page)?)
-                }
-                "save" => {
-                    let key = required_bounded_string(&input, "key", 1024, "storage")?;
-                    crate::storage::validate_storage_key(&key)?;
-                    let overwrite = input
-                        .get("overwrite")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    let source = input.get("source").ok_or_else(|| {
-                        BrainError::Invalid("storage save source is required".into())
-                    })?;
-                    match source.get("kind").and_then(serde_json::Value::as_str) {
-                        Some("inline_text") => {
-                            let text = required_bounded_string(
-                                source,
-                                "text",
-                                brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES,
-                                "storage save source",
-                            )?;
-                            if text.len() > brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES {
-                                return Err(BrainError::FileTooLarge {
-                                    limit: brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES,
-                                });
-                            }
-                            let object = write_storage_inline_state(
-                                self,
-                                session_id,
-                                st,
-                                key,
-                                base64::engine::general_purpose::STANDARD.encode(text.as_bytes()),
-                                Some("text/plain; charset=utf-8".into()),
-                                overwrite,
-                            )
-                            .await?;
-                            Ok(serde_json::to_value(object)?)
-                        }
-                        Some("sandbox_path") => {
-                            let path = required_bounded_string(
-                                source,
-                                "path",
-                                4096,
-                                "storage save source",
-                            )?;
-                            let generation =
-                                required_identifier(source, "generation", "storage save source")?;
-                            let target = default_sandbox_target(&st.head.root_id)?;
-                            let file_request = sandbox_file_request(&target, &generation, &path)?;
-                            let files = self.sandbox_files.as_ref().ok_or_else(|| {
-                                BrainError::Invalid("sandbox files are unavailable".into())
-                            })?;
-                            let entry = files
-                                .stat(file_request)
-                                .await
-                                .map_err(map_hand_port_error)?;
-                            if entry.kind != brain_protocol::hand::FileEntryKind::File {
-                                return Err(BrainError::Invalid(
-                                    "storage save source must be a regular file".into(),
-                                ));
-                            }
-                            let ticket = prepare_storage_upload_state(
-                                self,
-                                session_id,
-                                st,
-                                crate::storage::StorageUploadIntent {
-                                    key,
-                                    bytes: entry.bytes,
-                                    sha256: None,
-                                    content_type: None,
-                                    overwrite,
-                                },
-                            )
-                            .await?;
-                            let copy = sandbox_copy_request(
-                                operation_id,
-                                &target,
-                                &generation,
-                                &path,
-                                None,
-                                &ticket,
-                                "export",
-                                false,
-                            )?;
-                            let expected_digest = copy.request_digest.clone();
-                            let result = files.transfer(copy).await.map_err(map_hand_port_error)?;
-                            validate_sandbox_copy_result(&result, operation_id, &expected_digest)?;
-                            let exported = result.object.as_ref().ok_or_else(|| {
-                                BrainError::Hand(
-                                    "sandbox export omitted its uploaded object identity".into(),
-                                )
-                            })?;
-                            if exported.object_id.as_str() != ticket.object_id
-                                || exported.bytes != entry.bytes
-                            {
-                                return Err(BrainError::Hand(
-                                    "sandbox export returned a different object identity".into(),
-                                ));
-                            }
-                            let object = complete_storage_upload_state(
-                                self,
-                                session_id,
-                                st,
-                                ticket.transfer_id,
-                            )
-                            .await?;
-                            if object.bytes != exported.bytes
-                                || object.sha256 != exported.sha256.as_str()
-                            {
-                                return Err(BrainError::Journal(
-                                    "published storage object differs from the sandbox export"
-                                        .into(),
-                                ));
-                            }
-                            Ok(serde_json::to_value(object)?)
-                        }
-                        Some(other) => Err(BrainError::Invalid(format!(
-                            "unknown storage save source kind {other:?}"
-                        ))),
-                        None => Err(BrainError::Invalid(
-                            "storage save source kind is required".into(),
-                        )),
-                    }
-                }
-                "load" => {
-                    let key = required_bounded_string(&input, "key", 1024, "storage")?;
-                    crate::storage::validate_storage_key(&key)?;
-                    let path = required_bounded_string(&input, "path", 4096, "storage")?;
-                    let generation = required_identifier(&input, "generation", "storage")?;
-                    let overwrite = input
-                        .get("overwrite")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    ensure_storage_readable(&st.head, session_id)?;
-                    let storage = self.storage_port()?;
-                    let object = storage.stat(session_id, &key).await?;
-                    let ticket = storage.prepare_download(session_id, &key).await?;
-                    if ticket.max_bytes != object.bytes {
-                        return Err(BrainError::Journal(
-                            "download authority does not match the stored object size".into(),
-                        ));
-                    }
-                    let reference = storage_object_reference(
-                        &ticket.object_id,
-                        object.bytes,
-                        &object.sha256,
-                        object.content_type.as_deref(),
-                    )?;
-                    let target = default_sandbox_target(&st.head.root_id)?;
-                    let copy = sandbox_copy_request(
-                        operation_id,
-                        &target,
-                        &generation,
-                        &path,
-                        Some(reference),
-                        &ticket,
-                        "import",
-                        overwrite,
-                    )?;
-                    let expected_digest = copy.request_digest.clone();
-                    let result = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .transfer(copy)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    validate_sandbox_copy_result(&result, operation_id, &expected_digest)?;
-                    if result.object.is_some() {
-                        return Err(BrainError::Hand(
-                            "sandbox import returned an unexpected object identity".into(),
-                        ));
-                    }
-                    Ok(serde_json::to_value(result.file)?)
-                }
-                other => Err(BrainError::Invalid(format!(
-                    "unknown storage action {other:?}"
-                ))),
-            }
-        };
-        let result = tokio::select! {
-            result = operation => result,
-            _ = cancel.cancelled() => Err(BrainError::Cancelled),
-        };
-        engine_outcome(result, started)
-    }
-
     async fn persist_additional_sandbox_status(
         &self,
         current: &crate::journal::SandboxInventoryDoc,
@@ -3619,694 +2574,6 @@ impl Brain {
                 now_ms: crate::wall_ms(),
             })
             .await
-    }
-
-    async fn inspect_additional_sandbox(
-        &self,
-        current: crate::journal::SandboxInventoryDoc,
-    ) -> Result<crate::journal::SandboxInventoryDoc> {
-        if sandbox_status_releases_slot(&current.status) {
-            return Ok(current);
-        }
-        let control = self.sandbox_control.as_ref().ok_or_else(|| {
-            BrainError::Invalid("additional sandbox control is unavailable".into())
-        })?;
-        let status = match control.inspect(current.status.target.clone()).await {
-            Ok(status) => status,
-            Err(error) if error.code == brain_protocol::hand::HandErrorCode::SandboxGone => {
-                sandbox_gone_status(&current.status, "hand_reported_gone")?
-            }
-            Err(error) => return Err(map_hand_port_error(error)),
-        };
-        self.persist_additional_sandbox_status(&current, status)
-            .await
-    }
-
-    /// Every sandbox action requires the caller's observed generation: acting on a replaced
-    /// physical target is a conflict, never a silent redirect.
-    async fn additional_sandbox_for_action(
-        &self,
-        root_id: &str,
-        input: &serde_json::Value,
-    ) -> Result<(crate::journal::SandboxInventoryDoc, String)> {
-        let sandbox_id = required_identifier(input, "sandbox_id", "sandbox")?;
-        let mut item = self.journal.get_sandbox(root_id, &sandbox_id).await?;
-        if item.root_id != root_id {
-            return Err(BrainError::FileNotFound(format!("sandbox {sandbox_id}")));
-        }
-        let generation = required_identifier(input, "generation", "sandbox")?;
-        {
-            let observed = item
-                .status
-                .generation
-                .as_ref()
-                .map(|generation| generation.as_str());
-            if observed != Some(generation.as_str()) {
-                return Err(if sandbox_status_releases_slot(&item.status) {
-                    BrainError::SandboxGone
-                } else {
-                    BrainError::SandboxGenerationConflict
-                });
-            }
-        }
-        if sandbox_status_releases_slot(&item.status) {
-            return Err(BrainError::SandboxGone);
-        }
-        if item
-            .status
-            .expires_at_ms
-            .is_some_and(|expiry| expiry.get() <= crate::wall_ms())
-        {
-            item = self.inspect_additional_sandbox(item).await?;
-            if sandbox_status_releases_slot(&item.status) {
-                return Err(BrainError::SandboxGone);
-            }
-        }
-        if !matches!(
-            item.status.state,
-            brain_protocol::hand::SandboxState::Running
-                | brain_protocol::hand::SandboxState::Suspended
-        ) {
-            return Err(BrainError::HandUnavailable(
-                "additional sandbox has not reached a live state".into(),
-            ));
-        }
-        Ok((item, generation))
-    }
-
-    /// The closed official `brain.sandbox` capability. Logical inventory, authorization and
-    /// quota live in Brain; Hand receives only the exact typed target selected from that durable
-    /// inventory. No action can switch on a model-visible Tool name or fabricate a physical id.
-    pub(crate) async fn execute_sandbox_capability(
-        self: &Arc<Self>,
-        session_id: &str,
-        operation_id: &str,
-        input: serde_json::Value,
-        cancel: CancellationToken,
-        st: &mut TurnState,
-    ) -> Result<CallOutcome> {
-        let started = std::time::Instant::now();
-        let operation = async {
-            let action = input
-                .get("action")
-                .and_then(serde_json::Value::as_str)
-                .ok_or_else(|| BrainError::Invalid("sandbox action is required".into()))?;
-            let root_id = st.head.root_id.clone();
-            match action {
-                "create" => {
-                    let (sandbox_id, generation, target) =
-                        additional_sandbox_identity(&root_id, session_id, operation_id)?;
-                    let request_digest =
-                        sandbox_request_digest(&root_id, session_id, operation_id, &input)?;
-                    let now = crate::wall_ms();
-                    let creating: brain_protocol::hand::SandboxStatus =
-                        serde_json::from_value(serde_json::json!({
-                            "state": "creating",
-                            "target": target,
-                            "generation": generation,
-                            "changed_at_ms": now,
-                            "expires_at_ms": null,
-                        }))?;
-                    let reserved = self
-                        .journal
-                        .reserve_sandbox(&crate::journal::SandboxReserveRequest {
-                            root_id: root_id.clone(),
-                            owner_session_id: session_id.to_owned(),
-                            sandbox_id: sandbox_id.clone(),
-                            operation_id: operation_id.to_owned(),
-                            request_digest,
-                            generation_intent: generation.clone(),
-                            initial_status: creating,
-                            now_ms: now,
-                        })
-                        .await?;
-                    if sandbox_status_releases_slot(&reserved.status)
-                        || matches!(
-                            reserved.status.state,
-                            brain_protocol::hand::SandboxState::Running
-                                | brain_protocol::hand::SandboxState::Suspended
-                        )
-                    {
-                        return Ok(serde_json::json!({
-                            "sandbox_id": sandbox_id,
-                            "status": reserved.status,
-                        }));
-                    }
-                    let control = self.sandbox_control.as_ref().ok_or_else(|| {
-                        BrainError::Invalid("additional sandbox control is unavailable".into())
-                    })?;
-                    let request = sandbox_create_request(
-                        &st.head,
-                        reserved.status.target.clone(),
-                        &reserved.generation_intent,
-                    )?;
-                    let status = match control.create(request).await {
-                        Ok(status) => status,
-                        Err(error)
-                            if error.code == brain_protocol::hand::HandErrorCode::SandboxGone =>
-                        {
-                            sandbox_gone_status(&reserved.status, "hand_reported_gone")?
-                        }
-                        Err(error) => return Err(map_hand_port_error(error)),
-                    };
-                    let persisted = self
-                        .persist_additional_sandbox_status(&reserved, status)
-                        .await?;
-                    Ok(serde_json::json!({
-                        "sandbox_id": sandbox_id,
-                        "status": persisted.status,
-                    }))
-                }
-                "list" => {
-                    let cursor = optional_bounded_string(&input, "cursor", 4096)?;
-                    let limit = input
-                        .get("limit")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(20);
-                    if !(1..=100).contains(&limit) {
-                        return Err(BrainError::Invalid(
-                            "sandbox list limit must be between 1 and 100".into(),
-                        ));
-                    }
-                    let page = self
-                        .journal
-                        .list_sandbox_page(&crate::journal::SandboxListQuery {
-                            root_id: &root_id,
-                            limit: limit as usize,
-                            cursor: cursor.as_deref(),
-                        })
-                        .await?;
-                    let data = page
-                        .sandboxes
-                        .into_iter()
-                        .map(|item| {
-                            serde_json::json!({
-                                "sandbox_id": item.sandbox_id,
-                                "owner_session_id": item.owner_session_id,
-                                "status": item.status,
-                            })
-                        })
-                        .collect::<Vec<_>>();
-                    Ok(serde_json::json!({
-                        "has_more": page.next_cursor.is_some(),
-                        "next_cursor": page.next_cursor,
-                        "data": data,
-                    }))
-                }
-                "status" => {
-                    let sandbox_id = required_identifier(&input, "sandbox_id", "sandbox")?;
-                    let item = self.journal.get_sandbox(&root_id, &sandbox_id).await?;
-                    let item = self.inspect_additional_sandbox(item).await?;
-                    Ok(serde_json::json!({
-                        "sandbox_id": item.sandbox_id,
-                        "owner_session_id": item.owner_session_id,
-                        "status": item.status,
-                    }))
-                }
-                "exec" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let expected_target_ref = item
-                        .status
-                        .target_ref
-                        .as_ref()
-                        .map(|value| value.as_str().to_owned())
-                        .ok_or_else(|| {
-                            BrainError::Hand(
-                                "live sandbox status is missing its target reference".into(),
-                            )
-                        })?;
-                    let command = required_bounded_string(&input, "command", 131_072, "sandbox")?;
-                    let cwd = optional_bounded_string(&input, "cwd", 4096)?;
-                    let interactive = input
-                        .get("interactive")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    let execution_digest = hex::encode(Sha256::digest(
-                        format!("aex.sandbox-execution\0{root_id}\0{operation_id}").as_bytes(),
-                    ));
-                    let execution_id = format!("exe_{}", &execution_digest[..24]);
-                    let mut request: brain_protocol::hand::SandboxExecutionRequest =
-                        serde_json::from_value(serde_json::json!({
-                            "target": item.status.target,
-                            "expected_generation": generation,
-                            "execution_id": execution_id,
-                            "request_digest": "0".repeat(64),
-                            "input": {
-                                "command": command,
-                                "cwd": cwd,
-                                "interactive": interactive,
-                            },
-                            "resources": {
-                                "timeout_ms": 600_000,
-                                "max_output_bytes": brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES,
-                            },
-                            "network": sealed_sandbox_network(&st.head)?,
-                        }))?;
-                    request.request_digest =
-                        brain_protocol::contract::sandbox_execution_request_digest(&request);
-                    let expected_digest = String::from(request.request_digest.clone());
-                    let receipt = self
-                        .sandbox_control
-                        .as_ref()
-                        .ok_or_else(|| {
-                            BrainError::Invalid("additional sandbox control is unavailable".into())
-                        })?
-                        .execute(request)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    if String::from(receipt.operation.operation_id.clone()) != execution_id
-                        || String::from(receipt.operation.request_digest.clone()) != expected_digest
-                        || serde_json::to_value(&receipt.operation.target)?
-                            != serde_json::to_value(&item.status.target)?
-                        || String::from(receipt.operation.generation.clone()) != generation
-                        || String::from(receipt.operation.target_ref.clone()) != expected_target_ref
-                        || serde_json::to_value(&receipt.observation.operation)?
-                            != serde_json::to_value(&receipt.operation)?
-                    {
-                        return Err(BrainError::Hand(
-                            "sandbox execution receipt identity mismatch".into(),
-                        ));
-                    }
-                    if let Some(terminal) = &receipt.observation.terminal
-                        && (brain_protocol::contract::terminal_result_digest(terminal)
-                            != terminal.terminal_digest
-                            || terminal.inline.as_ref().is_some_and(|value| {
-                                !brain_protocol::contract::terminal_inline_fits(value)
-                            }))
-                    {
-                        return Err(BrainError::Hand(
-                            "sandbox execution terminal receipt is invalid or oversized".into(),
-                        ));
-                    }
-                    Ok(serde_json::json!({
-                        "execution_id": execution_id,
-                        "state": receipt.observation.state,
-                        "output": receipt.observation.output,
-                        "terminal": receipt.observation.terminal,
-                    }))
-                }
-                "write_stdin" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let expected_target_ref = item
-                        .status
-                        .target_ref
-                        .as_ref()
-                        .map(|value| value.as_str().to_owned())
-                        .ok_or_else(|| {
-                            BrainError::Hand(
-                                "live sandbox status is missing its target reference".into(),
-                            )
-                        })?;
-                    let eof = input
-                        .get("eof")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    let execution_id = required_identifier(&input, "execution_id", "sandbox")?;
-                    let text = optional_bounded_string(&input, "text", 4096)?.unwrap_or_default();
-                    if text.len() > 4096 {
-                        return Err(BrainError::Invalid(
-                            "sandbox write_stdin text exceeds 4096 UTF-8 bytes".into(),
-                        ));
-                    }
-                    let mut request: brain_protocol::hand::WriteStdinRequest =
-                        serde_json::from_value(serde_json::json!({
-                            "operation_id": operation_id,
-                            "request_digest": "0".repeat(64),
-                            "target": item.status.target,
-                            "expected_generation": generation,
-                            "execution_id": execution_id,
-                            "text": text,
-                            "eof": eof,
-                        }))?;
-                    request.request_digest =
-                        brain_protocol::contract::write_stdin_request_digest(&request);
-                    let expected_digest = request.request_digest.clone();
-                    let receipt = self
-                        .sandbox_control
-                        .as_ref()
-                        .ok_or_else(|| {
-                            BrainError::Invalid("additional sandbox control is unavailable".into())
-                        })?
-                        .write_stdin(request)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    if String::from(receipt.operation_id.clone()) != operation_id
-                        || receipt.request_digest != expected_digest
-                        || String::from(receipt.observation.operation.operation_id.clone())
-                            != execution_id
-                        || serde_json::to_value(&receipt.observation.operation.target)?
-                            != serde_json::to_value(&item.status.target)?
-                        || String::from(receipt.observation.operation.generation.clone())
-                            != generation
-                        || String::from(receipt.observation.operation.target_ref.clone())
-                            != expected_target_ref
-                    {
-                        return Err(BrainError::Hand(
-                            "sandbox stdin receipt identity mismatch".into(),
-                        ));
-                    }
-                    if let Some(terminal) = &receipt.observation.terminal
-                        && (brain_protocol::contract::terminal_result_digest(terminal)
-                            != terminal.terminal_digest
-                            || terminal.inline.as_ref().is_some_and(|value| {
-                                !brain_protocol::contract::terminal_inline_fits(value)
-                            }))
-                    {
-                        return Err(BrainError::Hand(
-                            "sandbox stdin terminal receipt is invalid or oversized".into(),
-                        ));
-                    }
-                    Ok(serde_json::json!({
-                        "accepted": receipt.accepted,
-                        "replayed": receipt.replayed,
-                        "state": receipt.observation.state,
-                        "output": receipt.observation.output,
-                        "terminal": receipt.observation.terminal,
-                    }))
-                }
-                "list_files" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let cursor = optional_bounded_string(&input, "cursor", 4096)?;
-                    let limit = input
-                        .get("limit")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(50);
-                    if !(1..=100).contains(&limit) {
-                        return Err(BrainError::Invalid(
-                            "sandbox file list limit must be between 1 and 100".into(),
-                        ));
-                    }
-                    let page = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .list(crate::hand::SandboxFileListRequest {
-                            target: item.status.target,
-                            expected_generation: generation,
-                            path,
-                            cursor,
-                            limit: limit as u32,
-                        })
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    Ok(serde_json::to_value(page)?)
-                }
-                "stat_file" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let entry = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .stat(sandbox_file_request(
-                            &item.status.target,
-                            &generation,
-                            &path,
-                        )?)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    Ok(serde_json::to_value(entry)?)
-                }
-                "read_file" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let content = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .read(sandbox_file_request(
-                            &item.status.target,
-                            &generation,
-                            &path,
-                        )?)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    let bytes = base64::engine::general_purpose::STANDARD
-                        .decode(&content.content_base64)
-                        .map_err(|_| BrainError::Hand("sandbox returned invalid base64".into()))?;
-                    if bytes.len() > brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES {
-                        return Err(BrainError::FileTooLarge {
-                            limit: brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES,
-                        });
-                    }
-                    let text = String::from_utf8(bytes).map_err(|_| {
-                        BrainError::Invalid(
-                            "sandbox read_file is model-inline UTF-8 only; use save for binary data"
-                                .into(),
-                        )
-                    })?;
-                    Ok(serde_json::json!({"entry": content.entry, "text": text}))
-                }
-                "write_file" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let text = required_bounded_string(
-                        &input,
-                        "text",
-                        brain_protocol::MAX_TOOL_TERMINAL_INLINE_BYTES,
-                        "sandbox",
-                    )?;
-                    let overwrite = input
-                        .get("overwrite")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    let request = sandbox_file_write_request(
-                        operation_id,
-                        &item.status.target,
-                        &generation,
-                        &path,
-                        text.as_bytes(),
-                        overwrite,
-                    )?;
-                    let expected_digest = request.request_digest.clone();
-                    let result = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .write(request)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    validate_sandbox_file_write_result(&result, operation_id, &expected_digest)?;
-                    Ok(serde_json::to_value(result.file)?)
-                }
-                "find_files" | "grep_files" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let field = if action == "find_files" {
-                        "glob"
-                    } else {
-                        "query"
-                    };
-                    let expression = required_bounded_string(&input, field, 4096, "sandbox")?;
-                    let cursor = optional_bounded_string(&input, "cursor", 4096)?;
-                    let limit = input
-                        .get("limit")
-                        .and_then(serde_json::Value::as_u64)
-                        .unwrap_or(50);
-                    if !(1..=100).contains(&limit) {
-                        return Err(BrainError::Invalid(
-                            "sandbox search limit must be between 1 and 100".into(),
-                        ));
-                    }
-                    let request = sandbox_search_request(
-                        &item.status.target,
-                        &generation,
-                        &path,
-                        &expression,
-                        cursor.as_deref(),
-                        limit as u32,
-                    )?;
-                    let files = self.sandbox_files.as_ref().ok_or_else(|| {
-                        BrainError::Invalid("sandbox files are unavailable".into())
-                    })?;
-                    let page = if action == "find_files" {
-                        files.find(request).await
-                    } else {
-                        files.grep(request).await
-                    }
-                    .map_err(map_hand_port_error)?;
-                    Ok(serde_json::to_value(page)?)
-                }
-                "load" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let key = required_bounded_string(&input, "key", 1024, "sandbox")?;
-                    crate::storage::validate_storage_key(&key)?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let overwrite = input
-                        .get("overwrite")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    ensure_storage_readable(&st.head, session_id)?;
-                    let storage = self.storage_port()?;
-                    let object = storage.stat(session_id, &key).await?;
-                    let ticket = storage.prepare_download(session_id, &key).await?;
-                    if ticket.max_bytes != object.bytes {
-                        return Err(BrainError::Journal(
-                            "download authority does not match the stored object size".into(),
-                        ));
-                    }
-                    let reference = storage_object_reference(
-                        &ticket.object_id,
-                        object.bytes,
-                        &object.sha256,
-                        object.content_type.as_deref(),
-                    )?;
-                    let copy = sandbox_copy_request(
-                        operation_id,
-                        &item.status.target,
-                        &generation,
-                        &path,
-                        Some(reference),
-                        &ticket,
-                        "import",
-                        overwrite,
-                    )?;
-                    let expected_digest = copy.request_digest.clone();
-                    let result = self
-                        .sandbox_files
-                        .as_ref()
-                        .ok_or_else(|| BrainError::Invalid("sandbox files are unavailable".into()))?
-                        .transfer(copy)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    validate_sandbox_copy_result(&result, operation_id, &expected_digest)?;
-                    if result.object.is_some() {
-                        return Err(BrainError::Hand(
-                            "sandbox import returned an unexpected object identity".into(),
-                        ));
-                    }
-                    Ok(serde_json::to_value(result.file)?)
-                }
-                "save" => {
-                    let (item, generation) =
-                        self.additional_sandbox_for_action(&root_id, &input).await?;
-                    let key = required_bounded_string(&input, "key", 1024, "sandbox")?;
-                    crate::storage::validate_storage_key(&key)?;
-                    let path = required_bounded_string(&input, "path", 4096, "sandbox")?;
-                    let overwrite = input
-                        .get("overwrite")
-                        .and_then(serde_json::Value::as_bool)
-                        .unwrap_or(false);
-                    let files = self.sandbox_files.as_ref().ok_or_else(|| {
-                        BrainError::Invalid("sandbox files are unavailable".into())
-                    })?;
-                    let entry = files
-                        .stat(sandbox_file_request(
-                            &item.status.target,
-                            &generation,
-                            &path,
-                        )?)
-                        .await
-                        .map_err(map_hand_port_error)?;
-                    if entry.kind != brain_protocol::hand::FileEntryKind::File {
-                        return Err(BrainError::Invalid(
-                            "sandbox save source must be a regular file".into(),
-                        ));
-                    }
-                    let ticket = prepare_storage_upload_state(
-                        self,
-                        session_id,
-                        st,
-                        crate::storage::StorageUploadIntent {
-                            key,
-                            bytes: entry.bytes,
-                            sha256: None,
-                            content_type: None,
-                            overwrite,
-                        },
-                    )
-                    .await?;
-                    let copy = sandbox_copy_request(
-                        operation_id,
-                        &item.status.target,
-                        &generation,
-                        &path,
-                        None,
-                        &ticket,
-                        "export",
-                        false,
-                    )?;
-                    let expected_digest = copy.request_digest.clone();
-                    let result = files.transfer(copy).await.map_err(map_hand_port_error)?;
-                    validate_sandbox_copy_result(&result, operation_id, &expected_digest)?;
-                    let exported = result.object.as_ref().ok_or_else(|| {
-                        BrainError::Hand("sandbox export omitted its object identity".into())
-                    })?;
-                    if exported.object_id.as_str() != ticket.object_id
-                        || exported.bytes != entry.bytes
-                    {
-                        return Err(BrainError::Hand(
-                            "sandbox export returned a different object identity".into(),
-                        ));
-                    }
-                    let object =
-                        complete_storage_upload_state(self, session_id, st, ticket.transfer_id)
-                            .await?;
-                    if object.bytes != exported.bytes || object.sha256 != exported.sha256.as_str() {
-                        return Err(BrainError::Journal(
-                            "published storage object differs from the sandbox export".into(),
-                        ));
-                    }
-                    Ok(serde_json::to_value(object)?)
-                }
-                "terminate" => {
-                    let sandbox_id = required_identifier(&input, "sandbox_id", "sandbox")?;
-                    let current = self.journal.get_sandbox(&root_id, &sandbox_id).await?;
-                    if sandbox_status_releases_slot(&current.status) {
-                        return Ok(serde_json::json!({
-                            "sandbox_id": sandbox_id,
-                            "status": current.status,
-                        }));
-                    }
-                    let status = match self
-                        .sandbox_control
-                        .as_ref()
-                        .ok_or_else(|| {
-                            BrainError::Invalid("additional sandbox control is unavailable".into())
-                        })?
-                        .terminate(current.status.target.clone())
-                        .await
-                    {
-                        Ok(status) => status,
-                        Err(error)
-                            if error.code == brain_protocol::hand::HandErrorCode::SandboxGone =>
-                        {
-                            sandbox_gone_status(&current.status, "hand_reported_gone")?
-                        }
-                        Err(error) => return Err(map_hand_port_error(error)),
-                    };
-                    if !sandbox_status_releases_slot(&status) {
-                        return Err(BrainError::Hand(
-                            "sandbox termination did not return a confirmed terminal state".into(),
-                        ));
-                    }
-                    let persisted = self
-                        .persist_additional_sandbox_status(&current, status)
-                        .await?;
-                    Ok(serde_json::json!({
-                        "sandbox_id": sandbox_id,
-                        "status": persisted.status,
-                    }))
-                }
-                other => Err(BrainError::Invalid(format!(
-                    "unknown sandbox action {other:?}"
-                ))),
-            }
-        };
-        let result = tokio::select! {
-            result = operation => result,
-            _ = cancel.cancelled() => Err(BrainError::Cancelled),
-        };
-        engine_outcome(result, started)
     }
 
     pub async fn delete(self: &Arc<Self>, session_id: &str) -> Result<()> {
@@ -5384,47 +3651,6 @@ fn replay_message(
     Ok(Some((replay.turn_id.clone(), replay.user_seq)))
 }
 
-#[derive(Debug)]
-struct PendingVolatile {
-    seq: u64,
-    turn: String,
-    agent: String,
-    call: String,
-    name: String,
-}
-
-#[derive(Debug, Clone)]
-struct PendingExternal {
-    seq: u64,
-    turn: String,
-    call: String,
-    name: String,
-    input: serde_json::Value,
-    context: HashMap<String, String>,
-    policy: crate::config::ServerToolPolicy,
-    parallel_batch: bool,
-}
-
-#[derive(Debug, Clone)]
-struct PendingCustomer {
-    seq: u64,
-    turn: String,
-    call: String,
-    name: String,
-    intent: crate::customer::CustomerOperationIntent,
-}
-
-#[derive(Debug, Clone)]
-struct PendingManaged {
-    seq: u64,
-    turn: String,
-    call: String,
-    name: String,
-    envelope: brain_protocol::hand::OperationEnvelope,
-    operation: Option<brain_protocol::hand::OperationRef>,
-    submit_unknown: bool,
-}
-
 fn resolve_sealed_tools(prefix: &PrefixDoc) -> Vec<crate::config::ToolDecl> {
     let mut tools = crate::tools::resolve(&prefix.tools).unwrap_or_default();
     for tool in &mut tools {
@@ -5438,354 +3664,6 @@ fn resolve_sealed_tools(prefix: &PrefixDoc) -> Vec<crate::config::ToolDecl> {
     tools
 }
 
-/// Host-tool calls whose intent committed but whose result did not. Only the stable sealed policy
-/// determines whether replay is permitted; model arguments cannot opt a call into replay.
-fn pending_external(entries: &[Entry], prefix: &PrefixDoc) -> Vec<PendingExternal> {
-    let answered: HashSet<&str> = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::ToolResult { call, .. } => Some(call.as_str()),
-            _ => None,
-        })
-        .collect();
-    let terminal_turns: HashSet<&str> = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::TurnCompleted { turn, .. } | Record::TurnFailed { turn, .. } => {
-                Some(turn.as_str())
-            }
-            _ => None,
-        })
-        .collect();
-    let contexts: HashMap<&str, &HashMap<String, String>> = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::UserMessage { turn, metadata, .. } => Some((turn.as_str(), metadata)),
-            _ => None,
-        })
-        .collect();
-    let policies: HashMap<String, crate::config::ServerToolPolicy> = resolve_sealed_tools(prefix)
-        .into_iter()
-        .filter_map(|tool| match tool.route {
-            crate::config::ToolRoute::Server(policy) => Some((tool.name, policy)),
-            _ => None,
-        })
-        .collect();
-
-    let mut pending = Vec::new();
-    for entry in entries {
-        let Record::ToolCall {
-            turn,
-            agent,
-            call,
-            name,
-            input,
-            ..
-        } = &entry.record
-        else {
-            continue;
-        };
-        let Some(policy) = policies.get(name.as_str()).cloned() else {
-            continue;
-        };
-        if agent != "root"
-            || answered.contains(call.as_str())
-            || terminal_turns.contains(turn.as_str())
-        {
-            continue;
-        }
-        let assistant_seq = entries
-            .iter()
-            .rev()
-            .find_map(|candidate| match &candidate.record {
-                Record::Assistant {
-                    turn: assistant_turn,
-                    agent,
-                    ..
-                } if candidate.seq < entry.seq && assistant_turn == turn && agent == "root" => {
-                    Some(candidate.seq)
-                }
-                _ => None,
-            })
-            .unwrap_or(0);
-        let next_assistant_seq = entries
-            .iter()
-            .filter_map(|candidate| match &candidate.record {
-                Record::Assistant {
-                    turn: assistant_turn,
-                    agent,
-                    ..
-                } if candidate.seq > assistant_seq && assistant_turn == turn && agent == "root" => {
-                    Some(candidate.seq)
-                }
-                _ => None,
-            })
-            .min()
-            .unwrap_or(u64::MAX);
-        let batch_size = entries
-            .iter()
-            .filter(|candidate| {
-                candidate.seq > assistant_seq
-                    && candidate.seq < next_assistant_seq
-                    && matches!(
-                        &candidate.record,
-                        Record::ToolCall { turn: other_turn, agent, .. }
-                            if other_turn == turn && agent == "root"
-                    )
-            })
-            .count();
-        pending.push(PendingExternal {
-            seq: entry.seq,
-            turn: turn.clone(),
-            call: call.clone(),
-            name: name.clone(),
-            input: input.clone(),
-            context: contexts
-                .get(turn.as_str())
-                .map_or_else(HashMap::new, |v| (*v).clone()),
-            policy,
-            parallel_batch: batch_size > 1,
-        });
-    }
-    pending.sort_by_key(|call| call.seq);
-    pending
-}
-
-fn pending_customer(
-    entries: &[Entry],
-    prefix: &PrefixDoc,
-    tenant_id: &str,
-    session_id: &str,
-) -> Vec<PendingCustomer> {
-    let answered = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::ToolResult { call, .. } => Some(call.as_str()),
-            _ => None,
-        })
-        .collect::<HashSet<_>>();
-    let intents = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::CustomerCallIntent {
-                call,
-                client_id,
-                process_id,
-                request_digest,
-                deadline_at_ms,
-                ..
-            } => Some((
-                call.as_str(),
-                (
-                    client_id.clone(),
-                    process_id.clone(),
-                    request_digest.clone(),
-                    *deadline_at_ms,
-                ),
-            )),
-            _ => None,
-        })
-        .collect::<HashMap<_, _>>();
-    let tools = resolve_sealed_tools(prefix)
-        .into_iter()
-        .map(|tool| (tool.name.clone(), tool))
-        .collect::<HashMap<_, _>>();
-    let mut pending = Vec::new();
-    for entry in entries {
-        let Record::ToolCall {
-            turn,
-            agent,
-            call,
-            name,
-            input,
-            ..
-        } = &entry.record
-        else {
-            continue;
-        };
-        if agent != "root" || answered.contains(call.as_str()) {
-            continue;
-        }
-        let Some((client_id, process_id, request_digest, deadline_at_ms)) =
-            intents.get(call.as_str())
-        else {
-            continue;
-        };
-        let Some(tool) = tools.get(name) else {
-            continue;
-        };
-        let crate::config::ToolRoute::Customer { registration } = &tool.route else {
-            continue;
-        };
-        pending.push(PendingCustomer {
-            seq: entry.seq,
-            turn: turn.clone(),
-            call: call.clone(),
-            name: name.clone(),
-            intent: crate::customer::CustomerOperationIntent {
-                tenant_id: tenant_id.to_owned(),
-                client_id: client_id.clone(),
-                process_id: process_id.clone(),
-                session_id: session_id.to_owned(),
-                operation_id: call.clone(),
-                registration: registration.clone(),
-                name: name.clone(),
-                contract_digest: tool.contract_digest.clone(),
-                input: input.clone(),
-                deadline_at_ms: *deadline_at_ms,
-                request_digest: request_digest.clone(),
-            },
-        });
-    }
-    pending.sort_by_key(|call| call.seq);
-    pending
-}
-
-fn pending_managed(entries: &[Entry]) -> Result<Vec<PendingManaged>> {
-    let mut pending = HashMap::<String, PendingManaged>::new();
-    for entry in entries {
-        match &entry.record {
-            Record::ManagedCallIntent {
-                turn,
-                call,
-                name,
-                envelope,
-            } => {
-                let next = PendingManaged {
-                    seq: entry.seq,
-                    turn: turn.clone(),
-                    call: call.clone(),
-                    name: name.clone(),
-                    envelope: envelope.clone(),
-                    operation: None,
-                    submit_unknown: false,
-                };
-                if let Some(previous) = pending.insert(call.clone(), next)
-                    && (previous.turn != *turn
-                        || previous.name != *name
-                        || serde_jcs::to_vec(&previous.envelope)? != serde_jcs::to_vec(envelope)?)
-                {
-                    return Err(BrainError::Journal(
-                        "managed operation id maps to conflicting durable intents".into(),
-                    ));
-                }
-            }
-            Record::ManagedCallAccepted {
-                turn,
-                call,
-                operation,
-            } => {
-                let item = pending.get_mut(call).ok_or_else(|| {
-                    BrainError::Journal(
-                        "managed accepted receipt has no preceding durable intent".into(),
-                    )
-                })?;
-                if item.turn != *turn {
-                    return Err(BrainError::Journal(
-                        "managed accepted receipt references a different turn".into(),
-                    ));
-                }
-                if let Some(previous) = &item.operation
-                    && serde_jcs::to_vec(previous)? != serde_jcs::to_vec(operation)?
-                {
-                    return Err(BrainError::Journal(
-                        "managed operation id maps to conflicting accepted receipts".into(),
-                    ));
-                }
-                item.operation = Some(operation.clone());
-            }
-            Record::ManagedCallUnknown {
-                turn,
-                call,
-                request_digest,
-            } => {
-                let item = pending.get_mut(call).ok_or_else(|| {
-                    BrainError::Journal(
-                        "managed unknown marker has no preceding durable intent".into(),
-                    )
-                })?;
-                if item.turn != *turn
-                    || item.envelope.request_digest.as_str() != request_digest.as_str()
-                {
-                    return Err(BrainError::Journal(
-                        "managed unknown marker conflicts with its durable intent".into(),
-                    ));
-                }
-                item.submit_unknown = true;
-            }
-            Record::ToolResult { call, .. } => {
-                pending.remove(call);
-            }
-            _ => {}
-        }
-    }
-    let mut pending = pending.into_values().collect::<Vec<_>>();
-    pending.sort_by_key(|call| call.seq);
-    Ok(pending)
-}
-
-/// Finds unanswered calls whose executor cannot be recovered unambiguously. A newly claimed
-/// session never reassigns these calls: Hand, customer-app and in-process intrinsic effects may
-/// already have happened. Replay-safe server capabilities are handled separately.
-fn pending_volatile(entries: &[Entry], prefix: &PrefixDoc) -> Vec<PendingVolatile> {
-    let customer_intents = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::CustomerCallIntent { call, .. } => Some(call.as_str()),
-            _ => None,
-        })
-        .collect::<HashSet<_>>();
-    let managed_intents = entries
-        .iter()
-        .filter_map(|entry| match &entry.record {
-            Record::ManagedCallIntent { call, .. } => Some(call.as_str()),
-            _ => None,
-        })
-        .collect::<HashSet<_>>();
-    let volatile_names: HashSet<String> = resolve_sealed_tools(prefix)
-        .into_iter()
-        .filter_map(|tool| match tool.route {
-            crate::config::ToolRoute::Server(_) => None,
-            _ => Some(tool.name),
-        })
-        .collect();
-    let mut pending = HashMap::<String, PendingVolatile>::new();
-    for entry in entries {
-        match &entry.record {
-            Record::ToolCall {
-                turn,
-                agent,
-                call,
-                name,
-                detach,
-                ..
-            } if volatile_names.contains(name)
-                && !customer_intents.contains(call.as_str())
-                && !managed_intents.contains(call.as_str())
-                && !detach =>
-            {
-                pending.insert(
-                    call.clone(),
-                    PendingVolatile {
-                        seq: entry.seq,
-                        turn: turn.clone(),
-                        agent: agent.clone(),
-                        call: call.clone(),
-                        name: name.clone(),
-                    },
-                );
-            }
-            Record::ToolResult { call, .. } => {
-                pending.remove(call);
-            }
-            _ => {}
-        }
-    }
-    let mut pending: Vec<_> = pending.into_values().collect();
-    pending.sort_by_key(|call| call.seq);
-    pending
-}
-
 struct RecoveredTurn {
     turn: String,
     context: HashMap<String, String>,
@@ -5793,678 +3671,7 @@ struct RecoveredTurn {
     tool_calls: u64,
 }
 
-async fn recover_customer_calls(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-    entries: &[Entry],
-) -> Result<bool> {
-    let Some(customer) = &brain.customer else {
-        if pending_customer(
-            entries,
-            &resident.st.head.prefix,
-            &resident.st.head.tenant_id,
-            session_id,
-        )
-        .is_empty()
-        {
-            return Ok(false);
-        }
-        return Err(BrainError::HandUnavailable(
-            "customer application coordinator is unavailable during recovery".into(),
-        ));
-    };
-    let pending = pending_customer(
-        entries,
-        &resident.st.head.prefix,
-        &resident.st.head.tenant_id,
-        session_id,
-    );
-    let mut changed = false;
-    for call in pending {
-        let execution = customer
-            .execute_prepared(
-                call.intent.clone(),
-                resident.st.head.prefix.customer_submit_retries,
-                CancellationToken::new(),
-            )
-            .await;
-        if execution.retryable_without_effect && crate::wall_ms() < call.intent.deadline_at_ms {
-            return Err(BrainError::HandUnavailable(
-                "sealed customer application process has not reconnected".into(),
-            ));
-        }
-        let tool = crate::tools::resolve(&resident.st.head.prefix.tools)?
-            .into_iter()
-            .find(|tool| tool.name == call.name);
-        let outcome = crate::tools::enforce_outcome(tool.as_ref(), &call.name, execution.outcome);
-        let content = if outcome.content.is_empty() {
-            format!("[{}: no output]", outcome.outcome)
-        } else {
-            outcome.content.clone()
-        };
-        let mut records = vec![(
-            resident.st.take_seq(),
-            Record::ToolResult {
-                turn: call.turn.clone(),
-                agent: "root".into(),
-                call: call.call.clone(),
-                name: call.name.clone(),
-                outcome: crate::events::tool_outcome(outcome.outcome),
-                content,
-                is_error: outcome.is_error,
-                exit_code: outcome.exit_code,
-                duration_ms: outcome.duration_ms,
-                truncated: outcome.truncated,
-            },
-        )];
-        if let Some(receipt) = &execution.terminal_receipt {
-            let ack = crate::journal::CustomerTerminalAckDoc {
-                turn: call.turn.clone(),
-                call: call.call.clone(),
-                client_id: call.intent.client_id.clone(),
-                process_id: receipt.process_id.clone(),
-                request_digest: receipt.request_digest.clone(),
-                terminal_digest: receipt.terminal_digest.clone(),
-            };
-            if !resident
-                .st
-                .head
-                .pending_customer_acks
-                .iter()
-                .any(|current| current.call == ack.call)
-            {
-                resident.st.head.pending_customer_acks.push(ack);
-            }
-            records.push((
-                resident.st.take_seq(),
-                Record::CustomerTerminalReceived {
-                    turn: call.turn.clone(),
-                    call: call.call.clone(),
-                    client_id: call.intent.client_id.clone(),
-                    process_id: receipt.process_id.clone(),
-                    request_digest: receipt.request_digest.clone(),
-                    terminal_digest: receipt.terminal_digest.clone(),
-                },
-            ));
-        }
-        commit(brain, session_id, &mut resident.st, records).await?;
-        changed = true;
-
-        if let Some(receipt) = execution.terminal_receipt
-            && customer.acknowledge_terminal(&receipt).await.is_ok()
-        {
-            let previous = resident.st.head.pending_customer_acks.clone();
-            resident.st.head.pending_customer_acks.retain(|pending| {
-                pending.call != receipt.operation_id
-                    || pending.request_digest != receipt.request_digest
-                    || pending.terminal_digest != receipt.terminal_digest
-            });
-            let acked = vec![(
-                resident.st.take_seq(),
-                Record::CustomerTerminalAcknowledged {
-                    turn: call.turn,
-                    call: call.call,
-                    request_digest: receipt.request_digest,
-                    terminal_digest: receipt.terminal_digest,
-                },
-            )];
-            if let Err(error) = commit(brain, session_id, &mut resident.st, acked).await {
-                resident.st.head.pending_customer_acks = previous;
-                return Err(error);
-            }
-        }
-    }
-    Ok(changed)
-}
-
-async fn recover_managed_calls(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-    entries: &[Entry],
-) -> Result<bool> {
-    use brain_protocol::hand::{HandErrorCode, OperationState};
-
-    let pending = pending_managed(entries)?;
-    if pending.is_empty() {
-        return Ok(false);
-    }
-    let tools = crate::tools::resolve(&resident.st.head.prefix.tools)?;
-    let active_turn = resident.st.head.turn.clone();
-    let rounds = resident.st.head.active_rounds;
-    let tool_calls = resident.st.head.active_tool_calls;
-    let mut recovered = Vec::new();
-    let mut sandbox_gone = None;
-
-    let (stale, pending): (Vec<_>, Vec<_>) = pending
-        .into_iter()
-        .partition(|call| active_turn.as_deref() != Some(call.turn.as_str()));
-    for call in stale {
-        if !call.submit_unknown {
-            let unknown = vec![(
-                resident.st.take_seq(),
-                Record::ManagedCallUnknown {
-                    turn: call.turn.clone(),
-                    call: call.call.clone(),
-                    request_digest: call.envelope.request_digest.to_string(),
-                },
-            )];
-            commit(brain, session_id, &mut resident.st, unknown).await?;
-        }
-        reconcile_managed_unknown_default_sandbox(brain, session_id, &mut resident.st).await?;
-        let outcome = crate::tools::enforce_outcome(
-            tools.iter().find(|tool| tool.name == call.name),
-            &call.name,
-            crate::turn::managed_unknown_call_outcome(&call.name),
-        );
-        let content = if outcome.content.is_empty() {
-            format!("[{}: no output]", outcome.outcome)
-        } else {
-            outcome.content
-        };
-        let result = vec![(
-            resident.st.take_seq(),
-            Record::ToolResult {
-                turn: call.turn,
-                agent: "root".into(),
-                call: call.call,
-                name: call.name,
-                outcome: crate::events::tool_outcome(outcome.outcome),
-                content,
-                is_error: outcome.is_error,
-                exit_code: outcome.exit_code,
-                duration_ms: outcome.duration_ms,
-                truncated: outcome.truncated,
-            },
-        )];
-        commit(brain, session_id, &mut resident.st, result).await?;
-    }
-    if pending.is_empty() {
-        return Ok(true);
-    }
-
-    let hand = brain.hand.as_ref().ok_or_else(|| {
-        BrainError::HandUnavailable(
-            "managed Hand is unavailable for durable operation recovery".into(),
-        )
-    })?;
-
-    'managed_calls: for call in pending {
-        if call.submit_unknown {
-            reconcile_managed_unknown_default_sandbox(brain, session_id, &mut resident.st).await?;
-            recovered.push((
-                call.clone(),
-                call.operation.clone(),
-                crate::turn::managed_unknown_call_outcome(&call.name),
-                None,
-            ));
-            continue;
-        }
-        let binding = resident.managed_bindings.get(&call.name).ok_or_else(|| {
-            BrainError::HandUnavailable(format!(
-                "managed Tool {} has no recovered immutable binding",
-                call.name
-            ))
-        })?;
-        if binding.binding_ref != call.envelope.binding_ref {
-            return Err(BrainError::Protocol(
-                "managed Tool binding changed across durable recovery".into(),
-            ));
-        }
-
-        let mut accepted_now = false;
-        let (operation, mut observation) = if let Some(operation) = call.operation.clone() {
-            crate::turn::verify_managed_operation(
-                &operation,
-                &call.call,
-                call.envelope.request_digest.as_str(),
-                session_id,
-                &resident.st.head,
-            )?;
-            let request: brain_protocol::hand::ObserveRequest =
-                serde_json::from_value(serde_json::json!({
-                    "operation": operation,
-                    "cursor": "",
-                    "wait_ms": binding.limits.max_wait_ms.min(30_000),
-                }))?;
-            let observed = tokio::time::timeout(
-                Duration::from_millis(request.wait_ms.saturating_add(1_000).max(1)),
-                hand.observe(request),
-            )
-            .await
-            .map_err(|_| {
-                BrainError::HandUnavailable("managed Tool recovery observation timed out".into())
-            })?;
-            match observed {
-                Ok(observation) => (operation, observation),
-                Err(error) if error.code == HandErrorCode::SandboxGone => {
-                    sandbox_gone = Some(managed_operation_gone_status(&operation)?);
-                    recovered.push((
-                        call,
-                        Some(operation),
-                        CallOutcome {
-                            outcome: TerminalOutcome::Interrupted,
-                            value: None,
-                            content: "managed Tool target disappeared before its durable terminal could be recovered".into(),
-                            is_error: true,
-                            exit_code: None,
-                            duration_ms: 0,
-                            truncated: false,
-                            terminal: None,
-                        },
-                        None,
-                    ));
-                    continue;
-                }
-                Err(error) => return Err(map_hand_port_error(error)),
-            }
-        } else {
-            let request = brain_protocol::hand::SubmitRequest {
-                envelope: call.envelope.clone(),
-                wait_up_to_ms: binding.limits.max_wait_ms.min(30_000),
-            };
-            let mut reprepared = false;
-            let receipt = loop {
-                let submitted = tokio::time::timeout(
-                    Duration::from_millis(request.wait_up_to_ms.saturating_add(1_000).max(1)),
-                    hand.submit(request.clone()),
-                )
-                .await
-                .map_err(|_| {
-                    BrainError::HandUnavailable("managed Tool recovery submit timed out".into())
-                })?;
-                match submitted {
-                    Ok(receipt) => break receipt,
-                    Err(error)
-                        if error.code == HandErrorCode::CapabilityUnavailable && !reprepared =>
-                    {
-                        let refreshed = brain
-                            .prepare_managed_session(session_id, &resident.st.head)
-                            .await?;
-                        let refreshed = refreshed.get(&call.name).ok_or_else(|| {
-                            BrainError::HandUnavailable(
-                                "managed Tool disappeared during exact re-preparation".into(),
-                            )
-                        })?;
-                        if refreshed.binding_ref != binding.binding_ref {
-                            return Err(BrainError::Protocol(
-                                "managed Tool binding changed during exact re-preparation".into(),
-                            ));
-                        }
-                        reprepared = true;
-                    }
-                    Err(error) if error.code == HandErrorCode::OperationUnknown => {
-                        let unknown = vec![(
-                            resident.st.take_seq(),
-                            Record::ManagedCallUnknown {
-                                turn: call.turn.clone(),
-                                call: call.call.clone(),
-                                request_digest: call.envelope.request_digest.to_string(),
-                            },
-                        )];
-                        commit(brain, session_id, &mut resident.st, unknown).await?;
-                        reconcile_managed_unknown_default_sandbox(
-                            brain,
-                            session_id,
-                            &mut resident.st,
-                        )
-                        .await?;
-                        recovered.push((
-                            call.clone(),
-                            None,
-                            crate::turn::managed_unknown_call_outcome(&call.name),
-                            None,
-                        ));
-                        continue 'managed_calls;
-                    }
-                    Err(error) => return Err(map_hand_port_error(error)),
-                }
-            };
-            crate::turn::verify_managed_operation(
-                &receipt.operation,
-                &call.call,
-                call.envelope.request_digest.as_str(),
-                session_id,
-                &resident.st.head,
-            )?;
-            crate::turn::verify_managed_observation(&receipt.observation, &receipt.operation)?;
-            accepted_now = true;
-            (receipt.operation, receipt.observation)
-        };
-
-        if accepted_now {
-            if let Some(target) = &observation.target {
-                resident.st.head.default_sandbox = Some(managed_operation_running_status(
-                    &operation,
-                    target.expires_at_ms,
-                )?);
-            }
-            let accepted = vec![(
-                resident.st.take_seq(),
-                Record::ManagedCallAccepted {
-                    turn: call.turn.clone(),
-                    call: call.call.clone(),
-                    operation: operation.clone(),
-                },
-            )];
-            commit(brain, session_id, &mut resident.st, accepted).await?;
-        }
-
-        crate::turn::verify_managed_observation(&observation, &operation)?;
-        if observation.state != OperationState::Terminal {
-            if crate::wall_ms() >= call.envelope.deadline_at_ms.get() {
-                let cancel: brain_protocol::hand::CancelRequest =
-                    serde_json::from_value(serde_json::json!({
-                        "operation": operation,
-                        "reason": "recovery_deadline_elapsed",
-                    }))?;
-                match hand.cancel(cancel).await {
-                    Ok(_) => {}
-                    Err(error) if error.code == HandErrorCode::SandboxGone => {
-                        sandbox_gone = Some(managed_operation_gone_status(&operation)?);
-                    }
-                    Err(error) => return Err(map_hand_port_error(error)),
-                }
-            }
-            let request: brain_protocol::hand::ObserveRequest =
-                serde_json::from_value(serde_json::json!({
-                    "operation": operation,
-                    "cursor": observation.next_cursor,
-                    "wait_ms": binding.limits.max_wait_ms.min(30_000),
-                }))?;
-            match tokio::time::timeout(
-                Duration::from_millis(request.wait_ms.saturating_add(1_000).max(1)),
-                hand.observe(request),
-            )
-            .await
-            .map_err(|_| {
-                BrainError::HandUnavailable("managed Tool recovery observation timed out".into())
-            })? {
-                Ok(next) => observation = next,
-                Err(error) if error.code == HandErrorCode::SandboxGone => {
-                    sandbox_gone = Some(managed_operation_gone_status(&operation)?);
-                    recovered.push((
-                        call,
-                        Some(operation),
-                        CallOutcome {
-                            outcome: TerminalOutcome::Interrupted,
-                            value: None,
-                            content: "managed Tool target disappeared before its durable terminal could be recovered".into(),
-                            is_error: true,
-                            exit_code: None,
-                            duration_ms: 0,
-                            truncated: false,
-                            terminal: None,
-                        },
-                        None,
-                    ));
-                    continue;
-                }
-                Err(error) => return Err(map_hand_port_error(error)),
-            }
-            crate::turn::verify_managed_observation(&observation, &operation)?;
-        }
-        if observation.state != OperationState::Terminal {
-            return Err(BrainError::HandUnavailable(
-                "managed Tool remains in progress during durable recovery".into(),
-            ));
-        }
-        let terminal = observation.terminal.ok_or_else(|| {
-            BrainError::Protocol(
-                "managed Hand reported terminal state without a terminal receipt".into(),
-            )
-        })?;
-        let (outcome, terminal_digest) = crate::turn::managed_terminal_call_outcome(terminal)?;
-        recovered.push((call, Some(operation), outcome, Some(terminal_digest)));
-    }
-
-    let mut records = Vec::new();
-    let mut terminal = None;
-    for (call, operation, outcome, terminal_digest) in recovered {
-        let tool = tools.iter().find(|tool| tool.name == call.name);
-        let outcome = crate::tools::enforce_outcome(tool, &call.name, outcome);
-        let content = if outcome.content.is_empty() {
-            format!("[{}: no output]", outcome.outcome)
-        } else {
-            outcome.content.clone()
-        };
-        records.push((
-            resident.st.take_seq(),
-            Record::ToolResult {
-                turn: call.turn.clone(),
-                agent: "root".into(),
-                call: call.call.clone(),
-                name: call.name.clone(),
-                outcome: crate::events::tool_outcome(outcome.outcome),
-                content,
-                is_error: outcome.is_error,
-                exit_code: outcome.exit_code,
-                duration_ms: outcome.duration_ms,
-                truncated: outcome.truncated,
-            },
-        ));
-        if let Some(terminal_digest) = terminal_digest {
-            let operation = operation.ok_or_else(|| {
-                BrainError::Journal(
-                    "managed terminal receipt has no accepted operation reference".into(),
-                )
-            })?;
-            let pending = crate::journal::ManagedTerminalAckDoc {
-                turn: call.turn.clone(),
-                call: call.call.clone(),
-                operation: operation.clone(),
-                terminal_digest: terminal_digest.clone(),
-            };
-            if !resident
-                .st
-                .head
-                .pending_managed_acks
-                .iter()
-                .any(|current| current.call == pending.call)
-            {
-                resident.st.head.pending_managed_acks.push(pending);
-            }
-            records.push((
-                resident.st.take_seq(),
-                Record::ManagedTerminalReceived {
-                    turn: call.turn.clone(),
-                    call: call.call.clone(),
-                    operation,
-                    terminal_digest,
-                },
-            ));
-        }
-        if terminal.is_none() {
-            terminal = outcome.terminal.map(|value| (call.call, call.name, value));
-        }
-    }
-    if let Some(status) = sandbox_gone {
-        resident.st.head.default_sandbox = Some(status.clone());
-        records.push((
-            resident.st.take_seq(),
-            Record::DefaultSandboxChanged { status },
-        ));
-    }
-    if let Some((call, name, terminal)) = terminal {
-        let turn = active_turn.expect("pending managed call required an active turn");
-        resident.st.head.state = resident.st.head.lifecycle_after_turn();
-        resident.st.head.turn = None;
-        resident.st.head.active_phase = None;
-        resident.st.head.provider_attempt = None;
-        resident.st.head.active_context.clear();
-        resident.st.head.active_rounds = 0;
-        resident.st.head.active_tool_calls = 0;
-        match terminal {
-            TurnTerminal::Complete { value, metadata } => records.push((
-                resident.st.take_seq(),
-                Record::TurnCompleted {
-                    turn: turn.clone(),
-                    stop_reason: TurnStopReason::EndTurn,
-                    rounds,
-                    tool_calls,
-                    result: Some(brain_protocol::session::TurnResult {
-                        call_id: call.parse().map_err(|error| {
-                            BrainError::Protocol(format!("managed call id: {error}"))
-                        })?,
-                        metadata,
-                        name,
-                        value,
-                    }),
-                },
-            )),
-            TurnTerminal::Fail { error } => records.push((
-                resident.st.take_seq(),
-                Record::TurnFailed {
-                    turn: turn.clone(),
-                    code: error.code.to_string(),
-                    message: error.message,
-                    details: error.details,
-                },
-            )),
-        }
-        records.push((
-            resident.st.take_seq(),
-            Record::State {
-                state: resident.st.head.state,
-                turn: None,
-            },
-        ));
-    } else {
-        resident.st.head.active_phase = Some(TurnPhase::ReadyToContinueModel);
-    }
-    commit(brain, session_id, &mut resident.st, records).await?;
-    let _ = recover_managed_terminal_acks(brain, session_id, resident).await;
-    Ok(true)
-}
-
 const MANAGED_UNKNOWN_SANDBOX_REASON: &str = "managed_operation_unknown_cleanup";
-
-/// Reconcile the rooted default target after Hand reports that Submit may have reached the guest
-/// but no operation receipt can be recovered. The durable `ManagedCallUnknown` marker is always
-/// committed by the caller first, so this routine may retry status/dematerialization freely but
-/// must never authorize another Submit.
-pub(crate) async fn reconcile_managed_unknown_default_sandbox(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    st: &mut TurnState,
-) -> Result<()> {
-    use brain_protocol::hand::{HandErrorCode, SandboxState};
-
-    let target = st
-        .head
-        .default_sandbox
-        .as_ref()
-        .map(|status| status.target.clone())
-        .unwrap_or(default_sandbox_target(&st.head.root_id)?);
-    let files = brain.sandbox_files.as_ref().ok_or_else(|| {
-        BrainError::HandUnavailable(
-            "managed Tool unknown-outcome status reconciliation is unavailable".into(),
-        )
-    })?;
-    let mut status = match files.status(target.clone()).await {
-        Ok(status) => status,
-        Err(error)
-            if matches!(
-                error.code,
-                HandErrorCode::SandboxGone | HandErrorCode::SandboxNotMaterialized
-            ) =>
-        {
-            let current = st
-                .head
-                .default_sandbox
-                .clone()
-                .unwrap_or(initial_default_sandbox(&st.head.root_id)?);
-            sandbox_gone_status(&current, "managed_operation_target_gone")?
-        }
-        Err(error) if error.retryable => {
-            return Err(BrainError::HandUnavailable(error.message.to_string()));
-        }
-        Err(error) => return Err(map_hand_port_error(error)),
-    };
-    if !sandbox_status_matches_target(&status, &target)? {
-        return Err(BrainError::Protocol(
-            "managed Tool unknown-outcome status references a different default target".into(),
-        ));
-    }
-    if matches!(status.state, SandboxState::NeverMaterialized) {
-        status = sandbox_gone_status(&status, "managed_operation_target_not_materialized")?;
-    }
-    if matches!(
-        status.state,
-        SandboxState::Running | SandboxState::Suspended
-    ) && (status.generation.is_none()
-        || status.target_ref.is_none()
-        || status.expires_at_ms.is_none())
-    {
-        return Err(BrainError::Protocol(
-            "managed Tool unknown-outcome status lacks generation, target_ref, or expiry".into(),
-        ));
-    }
-    if !sandbox_status_releases_slot(&status) {
-        let mut value = serde_json::to_value(&status)?;
-        value["reason"] = serde_json::Value::String(MANAGED_UNKNOWN_SANDBOX_REASON.into());
-        value["changed_at_ms"] = serde_json::Value::from(crate::wall_ms());
-        status = serde_json::from_value(value)?;
-    }
-    st.head.default_sandbox = Some(status.clone());
-    let seq = st.take_seq();
-    commit(
-        brain,
-        session_id,
-        st,
-        vec![(
-            seq,
-            Record::DefaultSandboxChanged {
-                status: status.clone(),
-            },
-        )],
-    )
-    .await?;
-    if sandbox_status_releases_slot(&status) {
-        return Ok(());
-    }
-
-    let preparation = brain.session_preparation.as_ref().ok_or_else(|| {
-        BrainError::HandUnavailable(
-            "managed Tool unknown-outcome dematerialization is unavailable".into(),
-        )
-    })?;
-    let terminal = match preparation.dematerialize_default(target.clone()).await {
-        Ok(status) => status,
-        Err(error)
-            if matches!(
-                error.code,
-                HandErrorCode::SandboxGone | HandErrorCode::SandboxNotMaterialized
-            ) =>
-        {
-            sandbox_gone_status(&status, "managed_operation_target_gone")?
-        }
-        Err(error) if error.retryable => {
-            return Err(BrainError::HandUnavailable(error.message.to_string()));
-        }
-        Err(error) => return Err(map_hand_port_error(error)),
-    };
-    if !sandbox_status_matches_target(&terminal, &target)? {
-        return Err(BrainError::Protocol(
-            "managed Tool unknown-outcome dematerialization references a different target".into(),
-        ));
-    }
-    if !sandbox_status_releases_slot(&terminal) {
-        return Err(BrainError::HandUnavailable(
-            "managed Tool unknown-outcome target has not reached a terminal sandbox state".into(),
-        ));
-    }
-    st.head.default_sandbox = Some(terminal.clone());
-    let seq = st.take_seq();
-    commit(
-        brain,
-        session_id,
-        st,
-        vec![(seq, Record::DefaultSandboxChanged { status: terminal })],
-    )
-    .await
-}
 
 fn managed_operation_running_status(
     operation: &brain_protocol::hand::OperationRef,
@@ -6494,299 +3701,6 @@ fn managed_operation_gone_status(
         "reason": "managed_operation_target_gone",
     }))
     .map_err(BrainError::from)
-}
-
-async fn recover_external_calls(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-    entries: &[Entry],
-) -> Result<Option<RecoveredTurn>> {
-    let Some(active_turn) = resident.st.head.turn.clone() else {
-        return Ok(None);
-    };
-    let pending: Vec<_> = pending_external(entries, &resident.st.head.prefix)
-        .into_iter()
-        .filter(|call| call.turn == active_turn)
-        .collect();
-    let rounds = resident.st.head.active_rounds;
-    let tool_calls = resident.st.head.active_tool_calls;
-    let context = resident.st.head.active_context.clone();
-    if pending.is_empty() {
-        return Ok(Some(RecoveredTurn {
-            turn: active_turn,
-            context,
-            rounds,
-            tool_calls,
-        }));
-    }
-    let sealed_tools = crate::tools::resolve(&resident.st.head.prefix.tools)?;
-    let mut records = Vec::with_capacity(pending.len() + 2);
-    let mut blocks = Vec::with_capacity(pending.len());
-    let mut terminal = None;
-    let mut unreplayable = false;
-
-    for call in pending {
-        let outcome =
-            if call.policy.effect == brain_protocol::session::ExternalToolEffect::ReplaySafe {
-                crate::turn::execute_external(
-                    brain.external_executor.clone(),
-                    call.policy,
-                    call.parallel_batch,
-                    session_id.to_string(),
-                    call.turn.clone(),
-                    "root".into(),
-                    call.call.clone(),
-                    call.name.clone(),
-                    call.input,
-                    call.context,
-                    CancellationToken::new(),
-                )
-                .await
-            } else {
-                unreplayable = true;
-                CallOutcome {
-                    outcome: TerminalOutcome::Interrupted,
-                    value: None,
-                    content: format!(
-                        "external tool {} was interrupted and its opaque effect was not replayed",
-                        call.name
-                    ),
-                    is_error: true,
-                    exit_code: None,
-                    duration_ms: 0,
-                    truncated: false,
-                    terminal: None,
-                }
-            };
-        let tool = sealed_tools.iter().find(|tool| tool.name == call.name);
-        let outcome = match tool {
-            Some(tool) => crate::tools::enforce_outcome(Some(tool), &call.name, outcome),
-            None => crate::tools::enforce_outcome(
-                None,
-                &call.name,
-                CallOutcome::failed(format!(
-                    "tool {} is absent from the recovered execution seal",
-                    call.name
-                )),
-            ),
-        };
-        let content = if outcome.content.is_empty() {
-            format!("[{}: no output]", outcome.outcome)
-        } else {
-            outcome.content.clone()
-        };
-        blocks.push(ContentBlock::ToolResult {
-            tool_use_id: call.call.clone(),
-            content: content.clone(),
-            is_error: outcome.is_error,
-        });
-        records.push((
-            resident.st.take_seq(),
-            Record::ToolResult {
-                turn: call.turn,
-                agent: "root".into(),
-                call: call.call.clone(),
-                name: call.name.clone(),
-                outcome: crate::events::tool_outcome(outcome.outcome),
-                content,
-                is_error: outcome.is_error,
-                exit_code: outcome.exit_code,
-                duration_ms: outcome.duration_ms,
-                truncated: outcome.truncated,
-            },
-        ));
-        if let Some(value) = outcome.terminal {
-            terminal = Some((call.call, call.name, value));
-        }
-    }
-
-    if let Some((call, name, terminal)) = terminal {
-        resident.st.head.state = resident.st.head.lifecycle_after_turn();
-        resident.st.head.turn = None;
-        resident.st.head.active_phase = None;
-        resident.st.head.provider_attempt = None;
-        resident.st.head.active_context.clear();
-        resident.st.head.active_rounds = 0;
-        resident.st.head.active_tool_calls = 0;
-        match terminal {
-            TurnTerminal::Complete { value, metadata } => {
-                records.push((
-                    resident.st.take_seq(),
-                    Record::TurnCompleted {
-                        turn: active_turn.clone(),
-                        stop_reason: TurnStopReason::EndTurn,
-                        rounds,
-                        tool_calls,
-                        result: Some(brain_protocol::session::TurnResult {
-                            call_id: call.parse().map_err(|error| {
-                                BrainError::Protocol(format!("external call id: {error}"))
-                            })?,
-                            metadata,
-                            name,
-                            value,
-                        }),
-                    },
-                ));
-            }
-            TurnTerminal::Fail { error } => records.push((
-                resident.st.take_seq(),
-                Record::TurnFailed {
-                    turn: active_turn.clone(),
-                    code: error.code.to_string(),
-                    message: error.message,
-                    details: error.details,
-                },
-            )),
-        }
-        records.push((
-            resident.st.take_seq(),
-            Record::State {
-                state: resident.st.head.state,
-                turn: None,
-            },
-        ));
-    } else if unreplayable {
-        resident.st.head.state = resident.st.head.lifecycle_after_turn();
-        resident.st.head.turn = None;
-        resident.st.head.active_phase = None;
-        resident.st.head.provider_attempt = None;
-        resident.st.head.active_context.clear();
-        resident.st.head.active_rounds = 0;
-        resident.st.head.active_tool_calls = 0;
-        records.push((
-            resident.st.take_seq(),
-            Record::TurnFailed {
-                turn: active_turn.clone(),
-                code: "cancelled".into(),
-                message:
-                    "turn interrupted with an opaque external-tool effect; it was not replayed"
-                        .into(),
-                details: None,
-            },
-        ));
-        records.push((
-            resident.st.take_seq(),
-            Record::State {
-                state: resident.st.head.state,
-                turn: None,
-            },
-        ));
-    }
-
-    commit(brain, session_id, &mut resident.st, records).await?;
-    resident.st.history.push(Message::tool_results(blocks));
-    if resident.st.head.turn.is_none() {
-        Ok(None)
-    } else {
-        Ok(Some(RecoveredTurn {
-            turn: active_turn,
-            context,
-            rounds,
-            tool_calls,
-        }))
-    }
-}
-
-/// Resolve the only ambiguous provider phase before a recovered turn is driven. Providers in the
-/// MVP do not expose a durable retrieval handle, so an unfinished intent becomes UNKNOWN. A new
-/// attempt is permitted only by the sealed crash-recovery budget; strict zero interrupts.
-async fn recover_provider_attempt(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-) -> Result<()> {
-    let Some(turn) = resident.st.head.turn.clone() else {
-        return Ok(());
-    };
-    let Some(mut attempt) = resident.st.head.provider_attempt.clone() else {
-        resident.st.head.active_phase = Some(TurnPhase::ReadyToBuildModelRequest);
-        commit(brain, session_id, &mut resident.st, vec![]).await?;
-        return Ok(());
-    };
-    let is_compaction = attempt.logical_operation_id.starts_with("cmp_");
-
-    let mut records = Vec::new();
-    if matches!(attempt.state.as_str(), "intent" | "running") {
-        records.push((
-            resident.st.take_seq(),
-            if is_compaction {
-                Record::CompactionUnknown {
-                    turn: turn.clone(),
-                    logical_operation_id: attempt.logical_operation_id.clone(),
-                    attempt_id: attempt.attempt_id.clone(),
-                    request_digest: attempt.request_digest.clone(),
-                    possibly_duplicated: true,
-                }
-            } else {
-                Record::ModelCallUnknown {
-                    turn: turn.clone(),
-                    logical_operation_id: attempt.logical_operation_id.clone(),
-                    attempt_id: attempt.attempt_id.clone(),
-                    request_digest: attempt.request_digest.clone(),
-                    possibly_duplicated: true,
-                }
-            },
-        ));
-        attempt.state = ProviderAttemptState::Unknown;
-    }
-    if attempt.state == ProviderAttemptState::ReplacementReady {
-        resident.st.head.provider_attempt = Some(attempt);
-        resident.st.head.active_phase = Some(if is_compaction {
-            TurnPhase::ReadyToCompact
-        } else {
-            TurnPhase::ReadyToBuildModelRequest
-        });
-        commit(brain, session_id, &mut resident.st, records).await?;
-        return Ok(());
-    }
-    if attempt.state != ProviderAttemptState::Unknown {
-        return Err(BrainError::Journal(format!(
-            "active provider attempt has invalid state {}",
-            attempt.state.as_str()
-        )));
-    }
-
-    if attempt.replacements_used < resident.st.head.prefix.provider_recovery_retries {
-        attempt.replacements_used += 1;
-        attempt.state = ProviderAttemptState::ReplacementReady;
-        resident.st.head.provider_attempt = Some(attempt);
-        resident.st.head.active_phase = Some(if is_compaction {
-            TurnPhase::ReadyToCompact
-        } else {
-            TurnPhase::ReadyToBuildModelRequest
-        });
-        commit(brain, session_id, &mut resident.st, records).await?;
-        return Ok(());
-    }
-
-    resident.st.head.state = resident.st.head.lifecycle_after_turn();
-    resident.st.head.turn = None;
-    resident.st.head.active_phase = None;
-    resident.st.head.provider_attempt = None;
-    resident.st.head.active_context.clear();
-    let rounds = resident.st.head.active_rounds;
-    let tool_calls = resident.st.head.active_tool_calls;
-    resident.st.head.active_rounds = 0;
-    resident.st.head.active_tool_calls = 0;
-    records.push((
-        resident.st.take_seq(),
-        Record::TurnCompleted {
-            turn: turn.clone(),
-            stop_reason: TurnStopReason::Interrupted,
-            rounds,
-            tool_calls,
-            result: None,
-        },
-    ));
-    records.push((
-        resident.st.take_seq(),
-        Record::State {
-            state: resident.st.head.state,
-            turn: None,
-        },
-    ));
-    commit(brain, session_id, &mut resident.st, records).await
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -7767,147 +4681,6 @@ async fn hydrate(brain: &Arc<Brain>, session_id: &str) -> Result<Resident> {
         .await;
     }
     Ok(resident)
-}
-
-async fn recover_customer_terminal_acks(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-) -> Result<()> {
-    let Some(customer) = &brain.customer else {
-        return if resident.st.head.pending_customer_acks.is_empty() {
-            Ok(())
-        } else {
-            Err(BrainError::HandUnavailable(
-                "customer coordinator is unavailable for a durable terminal acknowledgement".into(),
-            ))
-        };
-    };
-    let pending = resident.st.head.pending_customer_acks.clone();
-    let mut acknowledged = Vec::new();
-    for item in pending {
-        let receipt = crate::customer::CustomerTerminalReceipt {
-            operation_id: item.call.clone(),
-            request_digest: item.request_digest.clone(),
-            terminal_digest: item.terminal_digest.clone(),
-            process_id: item.process_id.clone(),
-        };
-        match customer
-            .acknowledge_durable_terminal(&resident.st.head.tenant_id, &item.client_id, &receipt)
-            .await
-        {
-            Ok(()) => acknowledged.push(item),
-            Err(error) => tracing::warn!(
-                session = session_id,
-                operation = %receipt.operation_id,
-                error = %error,
-                "durable customer terminal acknowledgement remains pending"
-            ),
-        }
-    }
-    if !acknowledged.is_empty() {
-        let records = acknowledged
-            .iter()
-            .map(|item| {
-                (
-                    resident.st.take_seq(),
-                    Record::CustomerTerminalAcknowledged {
-                        turn: item.turn.clone(),
-                        call: item.call.clone(),
-                        request_digest: item.request_digest.clone(),
-                        terminal_digest: item.terminal_digest.clone(),
-                    },
-                )
-            })
-            .collect::<Vec<_>>();
-        resident.st.head.pending_customer_acks.retain(|current| {
-            !acknowledged.iter().any(|item| {
-                current.call == item.call
-                    && current.request_digest == item.request_digest
-                    && current.terminal_digest == item.terminal_digest
-            })
-        });
-        commit(brain, session_id, &mut resident.st, records).await?;
-    }
-    if resident.st.head.pending_customer_acks.is_empty() {
-        Ok(())
-    } else {
-        Err(BrainError::HandUnavailable(
-            "customer terminal acknowledgement remains pending".into(),
-        ))
-    }
-}
-
-async fn recover_managed_terminal_acks(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Resident,
-) -> Result<()> {
-    let Some(hand) = &brain.hand else {
-        return if resident.st.head.pending_managed_acks.is_empty() {
-            Ok(())
-        } else {
-            Err(BrainError::HandUnavailable(
-                "managed Hand is unavailable for a durable terminal acknowledgement".into(),
-            ))
-        };
-    };
-    let pending = resident.st.head.pending_managed_acks.clone();
-    let mut acknowledged = Vec::new();
-    for item in pending {
-        let terminal_digest = item.terminal_digest.parse().map_err(|error| {
-            BrainError::Protocol(format!("persisted managed terminal digest: {error}"))
-        })?;
-        let request = brain_protocol::hand::AcknowledgeTerminalRequest {
-            operation: item.operation.clone(),
-            terminal_digest,
-        };
-        match hand.acknowledge_terminal(request).await {
-            Ok(ack) if ack.acknowledged => acknowledged.push(item),
-            Ok(_) => tracing::warn!(
-                session = session_id,
-                operation = item.operation.operation_id.as_str(),
-                "durable managed terminal acknowledgement was not accepted"
-            ),
-            Err(error) => tracing::warn!(
-                session = session_id,
-                operation = item.operation.operation_id.as_str(),
-                code = %error.code,
-                "durable managed terminal acknowledgement remains pending"
-            ),
-        }
-    }
-    if !acknowledged.is_empty() {
-        let records = acknowledged
-            .iter()
-            .map(|item| {
-                (
-                    resident.st.take_seq(),
-                    Record::ManagedTerminalAcknowledged {
-                        turn: item.turn.clone(),
-                        call: item.call.clone(),
-                        request_digest: item.operation.request_digest.to_string(),
-                        terminal_digest: item.terminal_digest.clone(),
-                    },
-                )
-            })
-            .collect::<Vec<_>>();
-        resident.st.head.pending_managed_acks.retain(|current| {
-            !acknowledged.iter().any(|item| {
-                current.operation.operation_id == item.operation.operation_id
-                    && current.operation.request_digest == item.operation.request_digest
-                    && current.terminal_digest == item.terminal_digest
-            })
-        });
-        commit(brain, session_id, &mut resident.st, records).await?;
-    }
-    if resident.st.head.pending_managed_acks.is_empty() {
-        Ok(())
-    } else {
-        Err(BrainError::HandUnavailable(
-            "managed terminal acknowledgement remains pending".into(),
-        ))
-    }
 }
 
 async fn commit(
@@ -9584,208 +6357,6 @@ fn matches_storage_publication(
         && object.publication_id.as_deref() == Some(upload.transfer_id.as_str())
 }
 
-async fn expire_storage_upload(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    st: &mut TurnState,
-) -> Result<()> {
-    let _heartbeat = start_lease_heartbeat(brain, session_id, &st.lease, true, None);
-    let Some(upload) = st.head.storage_upload.clone() else {
-        return Ok(());
-    };
-    if upload.state == UploadReservationState::Completed {
-        return Ok(());
-    }
-    let storage = brain.storage_port()?.clone();
-    if upload.state == UploadReservationState::Published {
-        let object = storage.stat(session_id, &upload.key).await?;
-        if !matches_storage_publication(&object, &upload) {
-            return Err(BrainError::Journal(
-                "published storage object does not match its durable reservation".into(),
-            ));
-        }
-        storage
-            .abort_upload(session_id, &upload.transfer_id)
-            .await?;
-        let mut completed = upload.clone();
-        completed.state = UploadReservationState::Completed;
-        st.head.storage_reserved_bytes = 0;
-        st.head.storage_upload = Some(completed);
-        let seq = st.take_seq();
-        return commit(
-            brain,
-            session_id,
-            st,
-            vec![(
-                seq,
-                Record::StorageUploadCompleted {
-                    transfer_id: upload.transfer_id,
-                    key: upload.key,
-                    bytes: upload.bytes,
-                    published_bytes: st.head.session_storage_bytes,
-                    reserved_bytes: st.head.storage_reserved_bytes,
-                },
-            )],
-        )
-        .await;
-    }
-    if upload.state == UploadReservationState::Reserved {
-        match storage.stat(session_id, &upload.key).await {
-            Ok(object) if matches_storage_publication(&object, &upload) => {
-                // `complete_upload` may have copied destination bytes before its response or the
-                // following journal decision was lost. Adopt that exact sealed result before
-                // touching staging, even after ticket expiry; otherwise visible bytes would be
-                // left unmetered while the reservation was released.
-                st.head.session_storage_bytes = st
-                    .head
-                    .session_storage_bytes
-                    .saturating_sub(upload.previous_bytes)
-                    .checked_add(upload.bytes)
-                    .ok_or_else(|| {
-                        BrainError::Journal("session storage meter overflowed".into())
-                    })?;
-                st.head.storage_reserved_bytes = 0;
-                let mut published = upload.clone();
-                published.sha256 = Some(object.sha256.clone());
-                published.state = UploadReservationState::Published;
-                st.head.storage_upload = Some(published);
-                let seq = st.take_seq();
-                commit(
-                    brain,
-                    session_id,
-                    st,
-                    vec![(
-                        seq,
-                        Record::StorageUploadPublished {
-                            transfer_id: upload.transfer_id.clone(),
-                            key: upload.key.clone(),
-                            bytes: upload.bytes,
-                            published_bytes: st.head.session_storage_bytes,
-                            reserved_bytes: st.head.storage_reserved_bytes,
-                        },
-                    )],
-                )
-                .await?;
-                // Publication is now durable. Staging cleanup may be retried independently and
-                // cannot make the published object or its tenant charge disappear.
-                storage
-                    .abort_upload(session_id, &upload.transfer_id)
-                    .await?;
-                let mut completed = upload.clone();
-                completed.sha256 = Some(object.sha256.clone());
-                completed.state = UploadReservationState::Completed;
-                st.head.storage_upload = Some(completed);
-                let seq = st.take_seq();
-                return commit(
-                    brain,
-                    session_id,
-                    st,
-                    vec![(
-                        seq,
-                        Record::StorageUploadCompleted {
-                            transfer_id: upload.transfer_id,
-                            key: upload.key,
-                            bytes: upload.bytes,
-                            published_bytes: st.head.session_storage_bytes,
-                            reserved_bytes: st.head.storage_reserved_bytes,
-                        },
-                    )],
-                )
-                .await;
-            }
-            Err(BrainError::FileNotFound(_)) => {}
-            // An overwrite may legitimately see the prior visible object before the staging
-            // copy runs. Equality of bytes/hash is not publication proof; only the intent id is.
-            Ok(_) if upload.overwrite => {}
-            Ok(_) => {
-                return Err(BrainError::Journal(
-                    "storage destination conflicts with its durable upload reservation".into(),
-                ));
-            }
-            Err(error) => return Err(error),
-        }
-    }
-    if upload.expires_at_ms > crate::wall_ms() {
-        return Ok(());
-    }
-    if upload.state == UploadReservationState::InlineReserved {
-        match storage.stat(session_id, &upload.key).await {
-            Ok(object) if matches_storage_publication(&object, &upload) => {
-                st.head.session_storage_bytes = st
-                    .head
-                    .session_storage_bytes
-                    .saturating_sub(upload.previous_bytes)
-                    .checked_add(upload.bytes)
-                    .ok_or_else(|| {
-                        BrainError::Journal("session storage meter overflowed".into())
-                    })?;
-                st.head.storage_reserved_bytes = 0;
-                let mut completed = upload.clone();
-                completed.state = UploadReservationState::Completed;
-                st.head.storage_upload = Some(completed);
-                let seq = st.take_seq();
-                return commit(
-                    brain,
-                    session_id,
-                    st,
-                    vec![(
-                        seq,
-                        Record::StorageUploadCompleted {
-                            transfer_id: upload.transfer_id,
-                            key: upload.key,
-                            bytes: upload.bytes,
-                            published_bytes: st.head.session_storage_bytes,
-                            reserved_bytes: st.head.storage_reserved_bytes,
-                        },
-                    )],
-                )
-                .await;
-            }
-            Err(BrainError::FileNotFound(_)) => {}
-            // The pre-existing overwrite target is not evidence that this inline intent ran.
-            // At expiry leave it untouched and release only the unpublished reservation.
-            Ok(_) if upload.overwrite => {}
-            Ok(_) => {
-                return Err(BrainError::Journal(
-                    "inline storage object conflicts with its durable intent".into(),
-                ));
-            }
-            Err(error) => return Err(error),
-        }
-    } else if upload.state != UploadReservationState::Reserved {
-        return Err(BrainError::Journal(format!(
-            "storage upload has invalid state {}",
-            upload.state.as_str()
-        )));
-    }
-    // Deletion precedes reservation release. A transient S3 failure therefore leaves the hard
-    // bound in place and the next operation retries cleanup instead of admitting more bytes.
-    if upload.state == UploadReservationState::Reserved {
-        storage
-            .abort_upload(session_id, &upload.transfer_id)
-            .await?;
-    }
-    st.head.storage_reserved_bytes = 0;
-    st.head.storage_upload = None;
-    let seq = st.take_seq();
-    commit(
-        brain,
-        session_id,
-        st,
-        vec![(
-            seq,
-            Record::StorageUploadExpired {
-                transfer_id: upload.transfer_id,
-                key: upload.key,
-                bytes: upload.bytes,
-                published_bytes: st.head.session_storage_bytes,
-                reserved_bytes: st.head.storage_reserved_bytes,
-            },
-        )],
-    )
-    .await
-}
-
 async fn reconcile_storage_delete(
     brain: &Arc<Brain>,
     session_id: &str,
@@ -9845,159 +6416,6 @@ async fn do_prepare_storage_upload(
 ) -> Result<crate::storage::StorageTransferTicket> {
     let r = ensure_resident(brain, session_id, resident).await?;
     prepare_storage_upload_state(brain, session_id, &mut r.st, intent).await
-}
-
-pub(crate) async fn prepare_storage_upload_state(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    st: &mut TurnState,
-    intent: crate::storage::StorageUploadIntent,
-) -> Result<crate::storage::StorageTransferTicket> {
-    prepare_storage_upload_state_for_transfer(brain, session_id, st, intent, None).await
-}
-
-async fn prepare_storage_upload_state_for_transfer(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    st: &mut TurnState,
-    intent: crate::storage::StorageUploadIntent,
-    requested_transfer_id: Option<String>,
-) -> Result<crate::storage::StorageTransferTicket> {
-    ensure_storage_readable(&st.head, session_id)?;
-    validate_storage_upload_intent(&intent, st.head.prefix.storage_max_object_bytes)?;
-    expire_storage_upload(brain, session_id, st).await?;
-
-    if let Some(upload) = &st.head.storage_upload {
-        let same_intent = upload.key == intent.key
-            && upload.bytes == intent.bytes
-            && (upload.sha256 == intent.sha256
-                || (requested_transfer_id.is_some() && intent.sha256.is_none()))
-            && upload.content_type == intent.content_type
-            && upload.overwrite == intent.overwrite;
-        let same_requested_transfer = requested_transfer_id
-            .as_deref()
-            .is_some_and(|transfer_id| transfer_id == upload.transfer_id);
-        if same_intent
-            && ((upload.state == UploadReservationState::Reserved
-                && (requested_transfer_id.is_none() || same_requested_transfer))
-                || (upload.state == UploadReservationState::Completed && same_requested_transfer))
-        {
-            return brain
-                .storage_port()?
-                .prepare_upload(crate::storage::StorageUploadRequest {
-                    session_id: session_id.to_owned(),
-                    transfer_id: upload.transfer_id.clone(),
-                    key: upload.key.clone(),
-                    bytes: upload.bytes,
-                    sha256: upload.sha256.clone(),
-                    content_type: upload.content_type.clone(),
-                    overwrite: upload.overwrite,
-                    expires_at_ms: upload.expires_at_ms,
-                })
-                .await;
-        }
-        if upload.state != UploadReservationState::Completed {
-            return Err(BrainError::StorageUploadInProgress {
-                transfer_id: upload.transfer_id.clone(),
-            });
-        }
-    }
-
-    let storage = brain.storage_port()?.clone();
-    let previous_bytes = match storage.stat(session_id, &intent.key).await {
-        Ok(object) if intent.overwrite => object.bytes,
-        Ok(_) => {
-            return Err(BrainError::Invalid(format!(
-                "session storage object {} already exists",
-                intent.key
-            )));
-        }
-        Err(BrainError::FileNotFound(_)) => 0,
-        Err(error) => return Err(error),
-    };
-    let visible_after_publish = st
-        .head
-        .session_storage_bytes
-        .saturating_sub(previous_bytes)
-        .checked_add(intent.bytes)
-        .ok_or_else(|| BrainError::Invalid("session storage byte count overflowed".into()))?;
-    // Completion briefly retains both the verified staging object and its published copy so a
-    // lost response remains retryable. Keep that worst-case physical footprint within quota.
-    let peak_bytes = visible_after_publish
-        .checked_add(intent.bytes)
-        .ok_or_else(|| BrainError::Invalid("session storage byte count overflowed".into()))?;
-    if peak_bytes > st.head.prefix.storage_max_session_bytes {
-        return Err(BrainError::StorageQuotaExceeded {
-            published: st.head.session_storage_bytes,
-            reserved: st.head.storage_reserved_bytes,
-            requested: intent.bytes,
-            limit: st.head.prefix.storage_max_session_bytes,
-        });
-    }
-
-    let transfer_id = requested_transfer_id.unwrap_or_else(|| crate::mint_id("xfer", 24));
-    let expires_at_ms = crate::wall_ms()
-        .checked_add(st.head.prefix.storage_transfer_ttl_ms)
-        .ok_or_else(|| BrainError::Invalid("storage transfer expiry overflowed".into()))?;
-    let upload = StorageUploadReservationDoc {
-        transfer_id: transfer_id.clone(),
-        key: intent.key.clone(),
-        bytes: intent.bytes,
-        sha256: intent.sha256.clone(),
-        content_type: intent.content_type.clone(),
-        overwrite: intent.overwrite,
-        previous_bytes,
-        expires_at_ms,
-        state: UploadReservationState::Reserved,
-    };
-    st.head.storage_reserved_bytes = intent.bytes;
-    st.head.storage_upload = Some(upload.clone());
-    let storage_gauges = (
-        st.head.session_storage_bytes,
-        st.head.storage_reserved_bytes,
-    );
-    let seq = st.take_seq();
-    commit(
-        brain,
-        session_id,
-        st,
-        vec![(
-            seq,
-            Record::StorageUploadReserved {
-                transfer_id: transfer_id.clone(),
-                key: intent.key.clone(),
-                bytes: intent.bytes,
-                sha256: intent.sha256.clone(),
-                expires_at_ms,
-                published_bytes: storage_gauges.0,
-                reserved_bytes: storage_gauges.1,
-            },
-        )],
-    )
-    .await?;
-
-    storage
-        .prepare_upload(crate::storage::StorageUploadRequest {
-            session_id: session_id.to_owned(),
-            transfer_id,
-            key: intent.key,
-            bytes: intent.bytes,
-            sha256: intent.sha256,
-            content_type: intent.content_type,
-            overwrite: intent.overwrite,
-            expires_at_ms,
-        })
-        .await
-}
-
-async fn do_complete_storage_upload(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Option<Resident>,
-    transfer_id: String,
-) -> Result<crate::storage::StorageObject> {
-    let r = ensure_resident(brain, session_id, resident).await?;
-    complete_storage_upload_state(brain, session_id, &mut r.st, transfer_id).await
 }
 
 pub(crate) async fn complete_storage_upload_state(
@@ -10113,28 +6531,6 @@ pub(crate) async fn complete_storage_upload_state(
     )
     .await?;
     Ok(object)
-}
-
-async fn do_write_storage_inline(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Option<Resident>,
-    key: String,
-    content_base64: String,
-    content_type: Option<String>,
-    overwrite: bool,
-) -> Result<crate::storage::StorageObject> {
-    let r = ensure_resident(brain, session_id, resident).await?;
-    write_storage_inline_state(
-        brain,
-        session_id,
-        &mut r.st,
-        key,
-        content_base64,
-        content_type,
-        overwrite,
-    )
-    .await
 }
 
 /// The turn path already owns the claimed session state. Reuse the same durable reservation and
@@ -10363,146 +6759,9 @@ pub(crate) async fn write_storage_inline_state(
     Ok(object)
 }
 
-async fn do_delete_storage_object(
-    brain: &Arc<Brain>,
-    session_id: &str,
-    resident: &mut Option<Resident>,
-    key: String,
-) -> Result<()> {
-    crate::storage::validate_storage_adapter_key(&key)?;
-    let r = ensure_resident(brain, session_id, resident).await?;
-    ensure_storage_readable(&r.st.head, session_id)?;
-    reconcile_storage_mutations(brain, session_id, &mut r.st).await?;
-    if let Some(upload) = &r.st.head.storage_upload
-        && upload.state != UploadReservationState::Completed
-    {
-        return Err(BrainError::StorageUploadInProgress {
-            transfer_id: upload.transfer_id.clone(),
-        });
-    }
-    if r.st.head.storage_delete.is_some() {
-        reconcile_storage_delete(brain, session_id, &mut r.st).await?;
-    }
-    let storage = brain.storage_port()?.clone();
-    let object = match storage.stat(session_id, &key).await {
-        Ok(object) => object,
-        Err(BrainError::FileNotFound(_)) => return Ok(()),
-        Err(error) => return Err(error),
-    };
-    let operation_id = crate::mint_id("del", 24);
-    r.st.head.storage_delete = Some(StorageDeleteReservationDoc {
-        operation_id: operation_id.clone(),
-        key: key.clone(),
-        bytes: object.bytes,
-        sha256: object.sha256.clone(),
-    });
-    let storage_gauges = (
-        r.st.head.session_storage_bytes,
-        r.st.head.storage_reserved_bytes,
-    );
-    let seq = r.st.take_seq();
-    commit(
-        brain,
-        session_id,
-        &mut r.st,
-        vec![(
-            seq,
-            Record::StorageDeleteIntent {
-                operation_id,
-                key,
-                bytes: object.bytes,
-                sha256: object.sha256,
-                published_bytes: storage_gauges.0,
-                reserved_bytes: storage_gauges.1,
-            },
-        )],
-    )
-    .await?;
-    reconcile_storage_delete(brain, session_id, &mut r.st).await
-}
-
 // ---------------------------------------------------------------------------------------------
 // Mapping helpers
 // ---------------------------------------------------------------------------------------------
-
-pub fn provider_name(p: &ApiProvider) -> &'static str {
-    match p {
-        ApiProvider::Openai => "openai",
-        ApiProvider::Anthropic => "anthropic",
-        ApiProvider::Deepseek => "deepseek",
-        ApiProvider::Moonshot => "moonshot",
-        ApiProvider::Xai => "xai",
-        ApiProvider::OpenaiCompatible => "openai_compatible",
-    }
-}
-
-/// Canonical public file path. The URL surface is deliberately narrower than hand tool paths:
-/// only absolute POSIX paths beneath `/workspace` are accepted.
-pub fn normalize_workspace_path(path: &str) -> Result<String> {
-    if path.len() > 4096 {
-        return Err(BrainError::Invalid(
-            "file path exceeds 4096 UTF-8 bytes".into(),
-        ));
-    }
-    if path.contains('\0') || path.contains('\\') {
-        return Err(BrainError::Invalid(
-            "file path contains a forbidden character".into(),
-        ));
-    }
-    let mut parts = path.split('/');
-    if parts.next() != Some("") || parts.next() != Some("workspace") {
-        return Err(BrainError::Invalid(
-            "file path must be absolute and beneath /workspace".into(),
-        ));
-    }
-    let mut clean = Vec::new();
-    for part in parts {
-        match part {
-            "" => continue,
-            "." | ".." => {
-                return Err(BrainError::Invalid(
-                    "file path may not contain . or .. components".into(),
-                ));
-            }
-            value => clean.push(value),
-        }
-    }
-    Ok(if clean.is_empty() {
-        "/workspace".into()
-    } else {
-        format!("/workspace/{}", clean.join("/"))
-    })
-}
-
-/// Certified: openai, anthropic. Available uncertified: the rest (they speak one of the two
-/// dialects). `openai_compatible` requires an explicit base_url.
-pub fn resolve_base_url(p: &ApiProvider, base_url: Option<&str>) -> Result<String> {
-    if let Some(u) = base_url {
-        if !u.starts_with("https://") {
-            return Err(BrainError::Invalid("model.base_url must be https".into()));
-        }
-        return Ok(u.trim_end_matches('/').to_string());
-    }
-    Ok(match p {
-        ApiProvider::Openai => "https://api.openai.com".into(),
-        ApiProvider::Anthropic => "https://api.anthropic.com".into(),
-        ApiProvider::Deepseek => "https://api.deepseek.com".into(),
-        ApiProvider::Moonshot => "https://api.moonshot.ai".into(),
-        ApiProvider::Xai => "https://api.x.ai".into(),
-        ApiProvider::OpenaiCompatible => {
-            return Err(BrainError::Invalid(
-                "model.base_url is required for provider openai_compatible".into(),
-            ));
-        }
-    })
-}
-
-pub fn dialect_of(provider: &str) -> Dialect {
-    match provider {
-        "anthropic" => Dialect::AnthropicMessages,
-        _ => Dialect::OpenAiChat,
-    }
-}
 
 /// Explicit Chat-Completions compatibility profile. OpenAI and generic compatible endpoints
 /// opt into the current OpenAI field; the named legacy-compatible providers retain the field
@@ -10515,312 +6774,12 @@ fn output_token_parameter(provider: &str) -> OutputTokenParameter {
     }
 }
 
-/// Rebuilds the sealed prefix from the HEAD prefix doc. Deterministic: the same doc always
-/// seals to the same digest.
-pub fn build_prefix(
-    p: &PrefixDoc,
-    max_rounds: u32,
-) -> Result<(crate::Shared<crate::config::SealedPrefix>, Dialect)> {
-    let dialect = dialect_of(&p.provider);
-    if dialect == Dialect::AnthropicMessages && p.reasoning_effort.is_some() {
-        return Err(BrainError::Invalid(
-            "model.reasoning_effort is not supported by the Anthropic MVP profile".into(),
-        ));
-    }
-    let mut decls = crate::tools::resolve(&p.tools)?;
-    for decl in &mut decls {
-        if let crate::config::ToolRoute::Intrinsic(capability) = &decl.route
-            && !crate::tools::is_direct_engine_capability(capability)
-        {
-            let policy = p
-                .official_capabilities
-                .get(capability)
-                .cloned()
-                .ok_or_else(|| {
-                    BrainError::Journal(format!(
-                        "sealed official capability {capability} has no trusted policy"
-                    ))
-                })?;
-            decl.route = crate::config::ToolRoute::Server(policy);
-        }
-    }
-    let mut def = AgentDef::new(
-        p.system_prompt
-            .clone()
-            .unwrap_or_else(default_system_prompt),
-        p.model.clone(),
-        dialect,
-    );
-    for d in decls {
-        def = def.tool(d);
-    }
-    def = def.sampling(GenOpts {
-        max_tokens: u32::try_from(p.max_output_tokens.unwrap_or(4096)).map_err(|_| {
-            BrainError::Journal("sealed max_output_tokens exceeds the canonical u32 bound".into())
-        })?,
-        output_token_parameter: output_token_parameter(&p.provider),
-        temperature: p.temperature.map(|t| t as f32),
-        reasoning_effort: p.reasoning_effort.clone(),
-        stop_sequences: Vec::new(),
-    });
-    def = def.limits(crate::config::Limits {
-        max_rounds,
-        ..crate::config::Limits::default()
-    });
-    let rendered_base = if p.rendered_base.is_null() {
-        None
-    } else {
-        let digest = hex::encode(Sha256::digest(serde_jcs::to_vec(&p.rendered_base)?));
-        if digest != p.rendered_base_digest {
-            return Err(BrainError::Journal(
-                "stored provider base segment digest does not match".into(),
-            ));
-        }
-        Some(p.rendered_base.clone())
-    };
-    let sealed = def.seal().with_provider_base(
-        rendered_base,
-        (!p.prompt_cache_key.is_empty()).then(|| p.prompt_cache_key.clone()),
-    );
-    Ok((sealed, dialect))
-}
-
 fn default_system_prompt() -> String {
     "You are an autonomous engineering agent running in an isolated Linux workspace \
      (/workspace, ARM64). Use the tools to inspect and change files and run commands. \
      Sandbox files are live compute state, not durable storage. Copy anything that must survive \
      target loss into session storage explicitly."
         .to_string()
-}
-
-/// Builds the contract Session document from the head. A sealed value that no longer parses
-/// as its contract type is journal corruption and errors loudly, naming the field — the REST
-/// read never substitutes placeholders or omits sealed identity.
-pub fn session_doc(session_id: &str, doc: &HeadDoc) -> Result<session::Session> {
-    let corrupt = |what: &str| {
-        BrainError::Journal(format!(
-            "session {session_id}: journaled {what} violates the public contract"
-        ))
-    };
-    let agentloop = Some(
-        match doc
-            .prefix
-            .agentloop
-            .clone()
-            .unwrap_or_else(crate::journal::AgentloopSelectorDoc::official_aex)
-        {
-            crate::journal::AgentloopSelectorDoc::Official { name, version } => {
-                session::AgentloopInfo::Official {
-                    name: name.parse().map_err(|_| corrupt("agentloop name"))?,
-                    version: version.parse().map_err(|_| corrupt("agentloop version"))?,
-                }
-            }
-            crate::journal::AgentloopSelectorDoc::Custom {
-                source_bundle_sha256,
-                toolchain,
-                ..
-            } => session::AgentloopInfo::Custom {
-                source_bundle_sha256: source_bundle_sha256
-                    .parse()
-                    .map_err(|_| corrupt("agentloop bundle digest"))?,
-                toolchain: toolchain
-                    .parse()
-                    .map_err(|_| corrupt("agentloop toolchain"))?,
-            },
-        },
-    );
-    Ok(session::Session {
-        agentloop,
-        context_fork: doc.context_fork.as_ref().map(public_context_fork),
-        created_at: crate::events::ts(doc.created_ms),
-        current_turn: doc
-            .turn
-            .as_deref()
-            .map(|t| t.parse().map_err(|_| corrupt("turn id")))
-            .transpose()?,
-        failure: doc.failure.as_ref().map(|f| session::SessionFailure {
-            at: crate::events::ts(f.at_ms),
-            code: match f.code.as_str() {
-                "binding_conflict" => session::SessionFailureCode::BindingConflict,
-                "provider_unusable" => session::SessionFailureCode::ProviderUnusable,
-                "hand_unavailable" => session::SessionFailureCode::HandUnavailable,
-                _ => session::SessionFailureCode::Internal,
-            },
-            message: f.message.clone(),
-        }),
-        id: session_id.parse().map_err(|_| corrupt("session id"))?,
-        parent_id: doc
-            .parent_id
-            .as_deref()
-            .map(|id| id.parse().map_err(|_| corrupt("parent session id")))
-            .transpose()?,
-        root_id: doc
-            .root_id
-            .parse()
-            .map_err(|_| corrupt("root session id"))?,
-        depth: i64::from(doc.depth),
-        last_seq: doc.last_seq,
-        last_message_at: doc.last_message_ms.map(crate::events::ts),
-        metadata: doc
-            .prefix
-            .metadata
-            .iter()
-            .map(|(key, value)| {
-                (
-                    key.as_str()
-                        .parse()
-                        .expect("sealed metadata key satisfies the public contract"),
-                    value
-                        .as_str()
-                        .parse()
-                        .expect("sealed metadata value satisfies the public contract"),
-                )
-            })
-            .collect(),
-        model: session::ModelInfo {
-            base_url: doc.prefix.base_url.clone(),
-            context_window_tokens: i64::from(doc.prefix.context_window_tokens),
-            name: doc.prefix.model.clone(),
-            provider: match doc.prefix.provider.as_str() {
-                "openai" => ApiProvider::Openai,
-                "anthropic" => ApiProvider::Anthropic,
-                "deepseek" => ApiProvider::Deepseek,
-                "moonshot" => ApiProvider::Moonshot,
-                "xai" => ApiProvider::Xai,
-                _ => ApiProvider::OpenaiCompatible,
-            },
-        },
-        name: doc
-            .child_name
-            .as_deref()
-            .map(|name| name.parse().map_err(|_| corrupt("child name")))
-            .transpose()?,
-        object: session::SessionObject::Session,
-        state: crate::events::session_state(doc.state),
-        turn_phase: doc
-            .active_phase
-            .map(|phase| phase.as_str().parse().map_err(|_| corrupt("turn phase")))
-            .transpose()?,
-        turn_state: crate::events::session_turn_state(doc.turn.as_deref()),
-        shape: doc.prefix.shape.clone(),
-        storage: session::StorageInfo {
-            session_storage_bytes: doc.session_storage_bytes,
-            upload_reserved_bytes: doc.storage_reserved_bytes,
-        },
-        turns: doc.turns,
-        updated_at: crate::events::ts(doc.updated_ms),
-    })
-}
-
-fn session_doc_summary(summary: &crate::journal::SessionSummary) -> Result<session::Session> {
-    let corrupt =
-        |what: &str| BrainError::Journal(format!("stored session summary has a corrupt {what}"));
-    Ok(session::Session {
-        // The bounded listing summary does not carry the sealed prefix; the full session
-        // resource does. Absent here, never a fabricated default.
-        agentloop: None,
-        context_fork: summary.context_fork.as_ref().map(public_context_fork),
-        created_at: crate::events::ts(summary.created_ms),
-        current_turn: summary.turn.as_deref().and_then(|turn| turn.parse().ok()),
-        failure: summary
-            .failure
-            .as_ref()
-            .map(|failure| session::SessionFailure {
-                at: crate::events::ts(failure.at_ms),
-                code: match failure.code.as_str() {
-                    "binding_conflict" => session::SessionFailureCode::BindingConflict,
-                    "provider_unusable" => session::SessionFailureCode::ProviderUnusable,
-                    "hand_unavailable" => session::SessionFailureCode::HandUnavailable,
-                    _ => session::SessionFailureCode::Internal,
-                },
-                message: failure.message.clone(),
-            }),
-        id: summary
-            .session_id
-            .parse()
-            .map_err(|_| corrupt("session id"))?,
-        parent_id: summary
-            .parent_id
-            .as_deref()
-            .map(|id| id.parse().map_err(|_| corrupt("parent id")))
-            .transpose()?,
-        root_id: summary.root_id.parse().map_err(|_| corrupt("root id"))?,
-        depth: i64::from(summary.depth),
-        last_seq: summary.last_seq,
-        last_message_at: summary.last_message_ms.map(crate::events::ts),
-        metadata: summary
-            .metadata
-            .iter()
-            .map(|(key, value)| {
-                (
-                    key.as_str()
-                        .parse()
-                        .expect("listed metadata key satisfies the public contract"),
-                    value
-                        .as_str()
-                        .parse()
-                        .expect("listed metadata value satisfies the public contract"),
-                )
-            })
-            .collect(),
-        model: session::ModelInfo {
-            base_url: summary.base_url.clone(),
-            context_window_tokens: i64::from(summary.context_window_tokens),
-            name: summary.model.clone(),
-            provider: match summary.provider.as_str() {
-                "openai" => ApiProvider::Openai,
-                "anthropic" => ApiProvider::Anthropic,
-                "deepseek" => ApiProvider::Deepseek,
-                "moonshot" => ApiProvider::Moonshot,
-                "xai" => ApiProvider::Xai,
-                _ => ApiProvider::OpenaiCompatible,
-            },
-        },
-        name: summary
-            .child_name
-            .as_deref()
-            .and_then(|name| name.parse().ok()),
-        object: session::SessionObject::Session,
-        state: crate::events::session_state(summary.state),
-        turn_phase: summary.active_phase.map(|phase| {
-            phase
-                .as_str()
-                .parse()
-                .expect("the protocol turn-phase vocabulary covers every journal phase")
-        }),
-        turn_state: crate::events::session_turn_state(summary.turn.as_deref()),
-        shape: summary.shape.clone(),
-        storage: session::StorageInfo {
-            session_storage_bytes: summary.session_storage_bytes,
-            upload_reserved_bytes: summary.storage_reserved_bytes,
-        },
-        turns: summary.turns,
-        updated_at: crate::events::ts(summary.updated_ms),
-    })
-}
-fn public_context_fork(fork: &ContextForkDoc) -> session::ContextFork {
-    session::ContextFork {
-        last_n: fork
-            .last_n
-            .and_then(|turns| std::num::NonZeroU64::new(u64::from(turns))),
-        mode: match fork.mode.as_str() {
-            "all" => session::ContextForkMode::All,
-            "none" => session::ContextForkMode::None,
-            "last_n" => session::ContextForkMode::LastN,
-            _ => panic!("sealed child context fork mode is a closed enum"),
-        },
-        resolved_turns: u64::from(fork.resolved_turns),
-        source_context_generation: fork.source_context_generation,
-        source_projection_digest: fork
-            .source_projection_digest
-            .parse()
-            .expect("sealed child context fork digest satisfies the public contract"),
-        source_session_id: fork
-            .source_session_id
-            .parse()
-            .expect("sealed child context fork source id satisfies the public contract"),
-        source_through_sequence: fork.source_through_sequence,
-    }
 }
 
 fn content_blocks(content: MessageRequestContent) -> Result<Vec<ContentBlock>> {
@@ -10853,5 +6812,5 @@ fn content_blocks(content: MessageRequestContent) -> Result<Vec<ContentBlock>> {
 }
 
 #[cfg(test)]
-#[path = "session_tests.rs"]
+#[path = "../session_tests.rs"]
 mod tests;
